% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0__globalVars.R
\docType{data}
\name{alphaNumeric}
\alias{alphaNumeric}
\title{Alpha-numeric characters}
\format{An object of class \code{character} of length 62.}
\usage{
alphaNumeric
}
\description{
Alpha-numeric characters
}
\keyword{datasets}

