% Generated by roxygen2 (4.0.2): do not edit by hand
\name{appendFromPlinkOut}
\alias{appendFromPlinkOut}
\title{Read a PLINK assoc file into a QCDH info hub}
\usage{
appendFromPlinkOut(hub, plinkOutFile)
}
\arguments{
\item{hub}{QCDH task info hub}

\item{plinkOutFile}{PLINK assoc filename}
}
\description{
Read a PLINK assoc file into a QCDH info hub
}

