% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\docType{methods}
\name{asBigMatrix}
\alias{asBigMatrix}
\alias{asBigMatrix,data.frame,ANY,ANY,ANY,ANY,ANY,ANY,ANY,logical-method}
\alias{asBigMatrix,matrix,ANY,ANY,ANY,ANY,ANY,ANY,ANY,logical-method}
\alias{asBigMatrix,vector,ANY,ANY,ANY,ANY,ANY,ANY,ANY,logical-method}
\title{Coerce an R vector/matrix/data.frame into a big.matrix}
\usage{
asBigMatrix(x, type = NULL, separated = FALSE, backingfile = NULL,
  backingpath = NULL, descriptorfile = NULL, binarydescriptor = FALSE,
  shared = TRUE, dimnames = FALSE)

\S4method{asBigMatrix}{matrix,ANY,ANY,ANY,ANY,ANY,ANY,ANY,logical}(x,
  type = NULL, separated = FALSE, backingfile = NULL,
  backingpath = NULL, descriptorfile = NULL, binarydescriptor = FALSE,
  shared = TRUE, dimnames = FALSE)

\S4method{asBigMatrix}{data.frame,ANY,ANY,ANY,ANY,ANY,ANY,ANY,logical}(x,
  type = NULL, separated = FALSE, backingfile = NULL,
  backingpath = NULL, descriptorfile = NULL, binarydescriptor = FALSE,
  shared = TRUE, dimnames = FALSE)

\S4method{asBigMatrix}{vector,ANY,ANY,ANY,ANY,ANY,ANY,ANY,logical}(x,
  type = NULL, separated = FALSE, backingfile = NULL,
  backingpath = NULL, descriptorfile = NULL, binarydescriptor = FALSE,
  shared = TRUE, dimnames = FALSE)
}
\arguments{
\item{x}{vector, matrix, or data.frmae}

\item{type}{See \code{bigmemory::as.big.matrix}}

\item{separated}{See \code{bigmemory::as.big.matrix}}

\item{backingfile}{See \code{bigmemory::as.big.matrix}}

\item{backingpath}{See \code{bigmemory::as.big.matrix}}

\item{descriptorfile}{See \code{bigmemory::as.big.matrix}}

\item{binarydescriptor}{See \code{bigmemory::as.big.matrix}}

\item{shared}{See \code{bigmemory::as.big.matrix}}

\item{dimnames}{logical. FALSE by default}
}
\value{
big.matrix object
}
\description{
This is a patched version of as.big.matrix from the bigmemory package.
The patch allows you to omit colnames/rownames even when they exist in
the R object.
}
\author{
Kaiyin Zhong, Fan Liu
}

