% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_filePath.R
\docType{methods}
\name{baseName}
\alias{baseName}
\alias{baseName,FilePath-method}
\title{Basename of a FilePath object}
\usage{
baseName(fp)

\S4method{baseName}{FilePath}(fp)
}
\arguments{
\item{fp}{character, file paths.}
}
\value{
character vector of basenames
}
\description{
Basename of a FilePath object
}
\examples{
\dontrun{
fp = filePath(R.home())
baseName(fp)
}


}
\author{
Kaiyin Zhong, Fan Liu
}

