% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bedcollinfo}
\alias{bedcollinfo}
\title{Collect info on shifted bed files}
\usage{
bedcollinfo(hub, genbed = FALSE, nShiftMin = 1, nShiftMax = 5)
}
\arguments{
\item{hub}{QCDH task info hub.}

\item{genbed}{Logical. Whether to generate shifted bed files.}

\item{nShiftMin}{Minimal shift number (only relevant when \code{genbed} is true).}

\item{nShiftMax}{Maximal shift number (only relevant when \code{genbed} is true).}
}
\description{
Collect info on shifted bed files, generate symlinks to bim and fam files, and put these info
in the QCDH task info hub.
}

