% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_bedcollr.R
\name{bedcollr}
\alias{bedcollr}
\title{Shift bed files}
\usage{
bedcollr(bfile = NULL, nshift_min = 1, nshift_max = NULL)
}
\arguments{
\item{bfile}{bed filename, without the \code{.bed} extension.}

\item{nshift_min}{Minimal shift number}

\item{nshift_max}{Maximal shift number}
}
\description{
This is a wrapper around the \code{bedcoll} commandline tool.
}

