% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{bedSizeCorrect}
\alias{bedSizeCorrect}
\title{Check whether bed file is of correct size}
\usage{
bedSizeCorrect(rbed_info)
}
\arguments{
\item{rbed_info}{RbedInfoC object}
}
\value{
logical.
}
\description{
It is correct if its real size is the equal to its theoretical size.
}
\author{
Kaiyin Zhong, Fan Liu
}

