% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{bimCorrectTypes}
\alias{bimCorrectTypes}
\title{Correct types of bim data.frame}
\usage{
bimCorrectTypes(bim_dat)
}
\arguments{
\item{bim_dat}{data.frame read from a .bim file}
}
\value{
data.frame
}
\description{
CHR, BP and GDIST columns should be integers.
}
\author{
Kaiyin Zhong, Fan Liu
}

