% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{bin2DescFilename}
\alias{bin2DescFilename}
\title{Convert a .bin filename to a .desc filename}
\usage{
bin2DescFilename(bin_file)
}
\arguments{
\item{bin_file}{character. .bin filename}
}
\value{
character
}
\description{
Convert a .bin filename to a .desc filename
}
\author{
Kaiyin Zhong, Fan Liu
}

