% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{binPhe}
\alias{binPhe}
\title{Check whether phenotype of a GWAS is binary}
\usage{
binPhe(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}
}
\value{
logical
}
\description{
Check whether phenotype of a GWAS is binary
}
\author{
Kaiyin Zhong, Fan Liu
}

