% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{bmAddCol}
\alias{bmAddCol}
\title{Add column(s) to an existing big.matrix}
\usage{
bmAddCol(bin_file, dat)
}
\arguments{
\item{bin_file}{character. Path to .bin file for file-backed big.matrix}

\item{dat}{vector, matrix or data.frame. Coercion rules are the same as in big.matrix}
}
\value{
updated description object.
}
\description{
This function provides an effecient way to append columns to a big.matrix (
without copying columns that are already on disk).
}
\author{
Kaiyin Zhong, Fan Liu
}

