% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{bmAttachBin}
\alias{bmAttachBin}
\title{Attach a big.matrix by its bin filename}
\usage{
bmAttachBin(bin_file)
}
\arguments{
\item{bin_file}{character. big.matrix bin filename}
}
\description{
Attach a big.matrix by its bin filename
}
\author{
Kaiyin Zhong
}

