% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{bmConvertFun}
\alias{bmConvertFun}
\title{Conversion function to use when appending values to a big.matrix}
\usage{
bmConvertFun(desc)
}
\arguments{
\item{desc}{description object}
}
\value{
conversion function.
}
\description{
Conversion function to use when appending values to a big.matrix
}
\author{
Kaiyin Zhong, Fan Liu
}

