% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{bmFilename}
\alias{bmFilename}
\title{Generate a big.matrix filename (.bin or .desc)}
\usage{
bmFilename(mat_name, type)
}
\arguments{
\item{mat_name}{character. Stem of filename.}

\item{type}{character. Either "bin" or "desc"}
}
\value{
character. big.matrix filename
}
\description{
Generate a big.matrix filename (.bin or .desc)
}
\author{
Kaiyin Zhong, Fan Liu
}

