% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{bmFilepath}
\alias{bmFilepath}
\title{Get the big.matrix file path according to GCDH task tag}
\usage{
bmFilepath(tag, mat_name, type)
}
\arguments{
\item{tag}{character. GCDH task tag.}

\item{mat_name}{character. nmiss, beta, stat, p, etc.}

\item{type}{character. Either "bin" or "desc"}
}
\description{
Get the big.matrix file path according to GCDH task tag
}
\author{
Kaiyin Zhong, Fan Liu
}

