% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{charify}
\alias{charify}
\title{Convert certain columns of a data.frame to character type}
\usage{
charify(dat, cols)
}
\arguments{
\item{dat}{data.frame}

\item{cols}{character. Names of columns to be converted.}
}
\value{
data.frame
}
\description{
Convert certain columns of a data.frame to character type
}
\examples{
\dontrun{
x = data.frame(x = 1:3, y= 2:4)
all(colClasses(x) == c("integer", "integer"))
x = charify(x, "x")
all(colClasses(x) == c("character", "integer"))
}

}
\author{
Kaiyin Zhong, Fan Liu
}

