% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkBinaryTrait}
\alias{checkBinaryTrait}
\title{Check whether a trait is binary}
\usage{
checkBinaryTrait(v, naVal = "NA")
}
\arguments{
\item{v}{A vector to be checked}

\item{naVal}{A value for NA coding}
}
\value{
A string. "bin" for binary, or "con" for continuous.
}
\description{
Check whether a trait is binary
}

