% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_nonExistentFiles.R
\name{checkFileExist}
\alias{checkFileExist}
\title{Stop when any file does not exist}
\usage{
checkFileExist(files)
}
\arguments{
\item{files}{character vector. File paths you want to check.}
}
\description{
Stop when any file does not exist
}
\author{
Kaiyin Zhong, Fan Liu
}

