% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{chExt}
\alias{chExt}
\title{Change extension names}
\usage{
chExt(filename, ext_name)
}
\arguments{
\item{filename}{character. File path}

\item{ext_name}{character. New extension name}
}
\description{
Change extension names
}
\author{
Kaiyin Zhong
}

