% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanByChrPvalsBonfer}
\alias{cleanByChrPvalsBonfer}
\title{Extract relevant SNP out of the QCDH data hub}
\usage{
cleanByChrPvalsBonfer(hub, threshold = 5e-08, lowerPbonfer = FALSE,
  forceRecalc = TRUE)
}
\arguments{
\item{hub}{QCDH data hub}

\item{threshold}{Numeric. Upper bound for minimal p value.}

\item{lowerPbonfer}{Logical. When this is set to TRUE, it requires that Bonferroni-corrected minimal P should be lower than p value of either of the paired SNPs. Default to FALSE.}

\item{forceRecalc}{Logical. When this is set to FALSE, this function will just return the cached info from \code{hub}. Default to TRUE.}
}
\value{
A data frame with info on all pairs of SNPs satisfying the aforementioned criteria.
}
\description{
Extract relevant SNP out of the QCDH data hub
}

