% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanSnpNames}
\alias{cleanSnpNames}
\title{Clean up SNP names for building linear models in R}
\usage{
cleanSnpNames(x)
}
\arguments{
\item{x}{A vector of SNP names to clean up.}
}
\value{
A cleanedup SNP names vector.
}
\description{
Some SNPs have not been give formal "rs..." names, for example "6:24331878:I".
These names start with a number and contain the special char ":", so they are not suitable
for usage in R formulas. This function cleans them up.
}

