% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{cmh}
\alias{cmh}
\title{Contrast Manhattan plot the simple way}
\usage{
cmh(gcdh_report, outfile = NULL)
}
\arguments{
\item{gcdh_report}{data.frame, from a GCDH analysis}

\item{outfile}{output image filepath. Any type (.png, .pdf, etc) supported by ggplot2::ggsave. Default to NULL. When it's not NULL, this function will try to save the plot to the specified path.}
}
\value{
A ggplot object
}
\description{
Contrast Manhattan plot the simple way
}
\author{
kaiyin
}

