% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{colClasses}
\alias{colClasses}
\title{Get classes of columns of a data.frame}
\usage{
colClasses(dat)
}
\arguments{
\item{dat}{data.frame}
}
\value{
character. Classes of \code{dat}
}
\description{
Get classes of columns of a data.frame
}
\examples{
\dontrun{
dat = data.frame(x = 15L, y = 3.14, z = "abc", 
  u = TRUE, stringsAsFactors = FALSE)
all(colClasses(dat) == 
				c("integer", "numeric", 
						"character", "logical"))
}

}
\author{
Kaiyin Zhong, Fan Liu
}

