% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{colCors}
\alias{colCors}
\title{Correlation coefficient of column-pairs of two data frames}
\usage{
colCors(dat1, dat2)
}
\arguments{
\item{dat1}{first data.frame}

\item{dat2}{second data.frame}
}
\value{
A vector of correlation coefficients.
}
\description{
Correlation coefficient of column-pairs of two data frames
}
\author{
Kaiyin Zhong
}

