% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collapse}
\alias{collapse}
\title{Genotype collapsing functions}
\usage{
collapse(gen_mat, n_shift)
}
\arguments{
\item{gen_mat}{Genotype matrix}

\item{n_shit}{Shift number.}
}
\value{
A matrix of collapsed genotypes.
}
\description{
See \url{https://gist.github.com/kindlychung/757f816f5b2f9e8d36b8#file-collapse-txt}
}

