% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{collapseMat}
\alias{collapseMat}
\title{Collapse two genotype matrices, column by column}
\usage{
collapseMat(m1, m2, collapse_matrix = matrix(c(0L, 0L, 0L, 0L, 0L, 1L, 1L, 1L,
  0L, 1L, 0L, 3L, 0L, 1L, 3L, 3L), 4, 4))
}
\arguments{
\item{m1}{first genotype matrix}

\item{m2}{second genotype matrix}

\item{collapse_matrix}{collapsed genotype matrix}
}
\value{
collapsed genotyp matrix
}
\description{
Each column is assummed to be the genotype for a SNP. The two genotype matrices should have the same size.
}
\author{
kaiyin
}

