% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{collClear}
\alias{collClear}
\title{Clear up CollapsABEL workspace}
\usage{
collClear()
}
\description{
The workspace folder is defined in \code{collenv$.collapsabel_dir}.
}
\author{
Kaiyin Zhong, Fan Liu
}

