% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0__globalVars.R
\docType{data}
\name{collenv}
\alias{collenv}
\title{An environment for storing CollapsABEL package local variables}
\format{An object of class \code{environment} of length 12.}
\usage{
collenv
}
\description{
.collapsabel_dir CollapsABEL home directory
}
\details{
.collapsabel_gwas CollapsABEL gwas directory

.collapsabel_gcdh CollapsABEL GCDH analysis directory

.assoc_header Plink .assoc file headers

.qassoc_header Plink .qassoc file headers

.logistic_header Plink .assoc.logistic file headers

.logistic_header_default Columns from plink .assoc.logistic file headers that are used by default

.linear_header Plink .assoc.linear file headers

.linear_header_default Columns from plink .assoc.linear file headers that are used by default

.plink_out_ext Plink output extensions

.plink_stdout Plink stdout

.plink_stderr Plink stderr
}
\keyword{datasets}

