% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{connectSnpPair}
\alias{connectSnpPair}
\title{Annotate a pair of SNPs in the contrast Manhattan plot}
\usage{
connectSnpPair(cplot, snp1, snp2, linetype = "dotted", hjust = 0,
  text_size = 3)
}
\arguments{
\item{cplot}{ggplot object. The contrast Manhattan plot to be annotated.}

\item{snp1}{character. First SNP.}

\item{snp2}{character. Second SNP.}

\item{linetype}{See \code{ggplot2::geom_segment}. Default to "dotted".}

\item{hjust}{See \code{ggplot2::annotate}. Default to 0.}

\item{text_size}{See \code{ggplot2::annotate}. Default to 3.}
}
\value{
ggplot object.
}
\description{
Annotate a pair of SNPs in the contrast Manhattan plot
}
\author{
Kaiyin Zhong
}

