% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{contrastData}
\alias{contrastData}
\title{Prepare data for \code{contrastPlot}}
\usage{
contrastData(chr, bp, p, gcdh_p, snp)
}
\arguments{
\item{chr}{integer. Chromosome vector.}

\item{bp}{integer. Position vector.}

\item{p}{numeric. P-value vector.}

\item{gcdh_p}{numeric. GCDH p-value vector.}

\item{snp}{character. SNP name vector.}
}
\description{
Prepare data for \code{contrastPlot}
}
\author{
Kaiyin Zhong
}

