% Generated by roxygen2 (4.0.2): do not edit by hand
\name{contrastPlot}
\alias{contrastPlot}
\title{Contrast QCDH p values with single-SNP p values}
\usage{
contrastPlot(hub, chrfilter = NULL, bplower = NULL, bpupper = NULL,
  pvallower = NULL, pvalupper = NULL, bonferroni = FALSE)
}
\arguments{
\item{hub}{QCDH data hub}

\item{chrfilter}{Numeric. A vector of chromosome numbers. If present, then only SNPs in these chromosomes are selected. Default to NULL.}

\item{bplower}{Numeric. If present, then only SNPs with base position greater than it will selected. Default to NULL.}

\item{bpupper}{Numeric. If present, then only SNPs with base position less than it will selected. Default to NULL.}

\item{pvallower}{Numeric. If present, then only SNPs with p values higher than it will be plotted in the single-SNP part. Default to NULL.}

\item{pvalupper}{Numeric. If present, then only SNPs with p values lower than it will be plotted in the single-SNP part. Default to NULL.}

\item{bonferroni}{Logical. If present, then the QCDH p values will be adjusted by Bonferroni correction.  Default to NULL.}
}
\value{
A ggplot object of contrast Manhattan plot. (This plot is also stored in the QCDH data hub.)
}
\description{
Contrast QCDH p values with single-SNP p values
}

