% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{correctDesc}
\alias{correctDesc}
\title{Correct description of big.matrix}
\usage{
correctDesc(desc_file)
}
\arguments{
\item{desc_file}{character. Path to description file}
}
\value{
list. Corrected description object.
}
\description{
Correct description of big.matrix
}
\author{
Kaiyin Zhong, Fan Liu
}

