% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\docType{methods}
\name{covarNames}
\alias{covarNames}
\title{Get covariate names of a GWAS}
\usage{
covarNames(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}
}
\value{
character. Vector of covariate names.
}
\description{
Get covariate names of a GWAS
}
\author{
Kaiyin Zhong, Fan Liu
}

