% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_cytoband.R
\name{cytoband}
\alias{cytoband}
\title{Find cytoband at a given position}
\usage{
cytoband(chr, pos, ref = "hg19")
}
\arguments{
\item{chr}{integer or character. Chromosome number. If it's an integer it should be in range [1, 22]. If it's a string it's should be in the format as "chr1, chr2, ..., chr22, chrX, chrY"}

\item{pos}{integer. Position on chromosome.}

\item{ref}{character. Reference data. Should be either "hg18" or "hg19"}
}
\value{
Vector of cytobands.
}
\description{
Find cytoband at a given position
}
\author{
kaiyin
}

