% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{datToVec}
\alias{datToVec}
\title{Extract one row or column of a data frame as a vector}
\usage{
datToVec(dat, i, row = TRUE)
}
\arguments{
\item{dat}{data.frame}

\item{i}{row or column number}

\item{row}{Logical. If TRUE, then i is the row number, otherwise i is the column number}
}
\value{
A vector.
}
\description{
Extract one row or column of a data frame as a vector
}
\author{
kaiyin
}

