% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbgtrigger}
\alias{dbgtrigger}
\title{Getting or setting a local boolean variable for debugging purposes}
\usage{
dbgtrigger(bool = NULL)
}
\arguments{
\item{bool}{Logical value that you want to set. When not provided, return the value previously set (default to FALSE)}
}
\description{
Getting or setting a local boolean variable for debugging purposes, thanks to flodel on stackoverflow
}
\examples{
dbgtrigger()
dbgtrigger(TRUE)
dbgtrigger()
}

