% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{desc2BinFilename}
\alias{desc2BinFilename}
\title{Convert a .desc filename to a .bin filename}
\usage{
desc2BinFilename(desc_file)
}
\arguments{
\item{desc_file}{character. .desc filename}
}
\value{
character
}
\description{
Convert a .desc filename to a .bin filename
}
\author{
Kaiyin Zhong, Fan Liu
}

