% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{dir.create2}
\alias{dir.create2}
\title{Create directory if it does not already exist}
\usage{
dir.create2(dir)
}
\arguments{
\item{dir}{character. Path of directory to be created.}
}
\description{
Create directory if it does not already exist
}
\author{
Kaiyin Zhong, Fan Liu
}

