% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_filePath.R
\docType{methods}
\name{dirName}
\alias{dirName}
\alias{dirName,FilePath-method}
\title{Directory name of a file path}
\usage{
dirName(fp)

\S4method{dirName}{FilePath}(fp)
}
\arguments{
\item{fp}{FilePath object}
}
\value{
character vector of directories
}
\description{
Directory name of a file path
}
\examples{
\dontrun{
fp = filePath(R.home())
dirName(fp)
}

}
\author{
Kaiyin Zhong, Fan Liu
}

