% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{evalFile}
\alias{evalFile}
\title{Eval R expressions from a file.}
\usage{
evalFile(filename)
}
\arguments{
\item{filename}{character}
}
\description{
Eval R expressions from a file.
}
\author{
Kaiyin Zhong, Fan Liu
}

