% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{famCorrectTypes}
\alias{famCorrectTypes}
\title{Correct types of fam data.frame}
\usage{
famCorrectTypes(fam_dat)
}
\arguments{
\item{fam_dat}{data.frame read from a .fam file}
}
\value{
data.frame
}
\description{
SEX and PHE columns should be integers.
}
\author{
Kaiyin Zhong, Fan Liu
}

