% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{file.create2}
\alias{file.create2}
\title{Create file if it does not already exist}
\usage{
file.create2(filename)
}
\arguments{
\item{filename}{character. Path of file to be created.}
}
\description{
Create file if it does not already exist
}
\author{
Kaiyin Zhong, Fan Liu
}

