% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_filePath.R
\docType{class}
\name{FilePath-class}
\alias{.FilePath}
\alias{FilePath-class}
\title{An S4 class to represent a file path}
\description{
This class comes with a validation function, 
making sure that the file exists.
}
\section{Slots}{

\describe{
\item{\code{path}}{character, file or dir path}
}}
\author{
Kaiyin Zhong, Fan Liu
}

