% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_filePath.R
\name{filePath}
\alias{filePath}
\title{Constructor for FilePath class}
\usage{
filePath(s)
}
\arguments{
\item{s}{character, path to file or dir}
}
\value{
FilePath object
}
\description{
Constructor for FilePath class
}
\examples{
\dontrun{
fp = filePath(R.home())
dirName(fp) == dirname(fp@path)
baseName(fp) == basename(fp@path)
}

}
\author{
Kaiyin Zhong, Fan Liu
}

