% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{fileSize}
\alias{fileSize}
\title{Get file size}
\usage{
fileSize(filename)
}
\arguments{
\item{filename}{character. Path to file.}
}
\value{
integer. Size of file.
}
\description{
Get file size
}
\author{
Kaiyin Zhong, Fan Liu
}

