% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{gcdhBmCreate}
\alias{gcdhBmCreate}
\title{Create a big.matrix under specified GCDH tag}
\usage{
gcdhBmCreate(tag, bm_name, nrow, ncol = 1)
}
\arguments{
\item{tag}{character. GCDH tag.}

\item{bm_name}{character. Name of the big.matrix to be created.}

\item{nrow}{integer. Number of rows of the big.matrix}

\item{ncol}{integer. Number of columns of the big.matrix. Default to 1.}
}
\description{
Create a big.matrix under specified GCDH tag
}
\author{
Kaiyin Zhong, Fan Liu
}

