% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{gcdhDir}
\alias{gcdhDir}
\title{Create GCDH task directories by tag}
\usage{
gcdhDir(gcdh_tag)
}
\arguments{
\item{gcdh_tag}{character. Tag for GCDH task.}
}
\value{
character. Directory of the task.
}
\description{
The task folder is a subfolder of the value of \code{collenv$.collapsabel_gcdh}.
It will be created if it does not yet exist.
}
\author{
Kaiyin Zhong, Fan Liu
}

