% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{gcdhReport}
\alias{gcdhReport}
\title{Generate a report from a GCDH run}
\usage{
gcdhReport(run_res)
}
\arguments{
\item{run_res}{Result from \code{runGcdh}}
}
\value{
path to SQLite database
}
\description{
For each p-value from a GCDH run, search for indices of the corresponding SNP pair.
Combine statistics from single-SNP approach with GCDH statistics.
}
\author{
Kaiyin Zhong, Fan Liu
}

