% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_haplo.R
\name{getHaplo}
\alias{getHaplo}
\title{Infer haplotypes for a pair of SNPs}
\usage{
getHaplo(geno, format_idx = NULL)
}
\arguments{
\item{geno}{Genotype data frame. Must have 4 columns, the first two being "FID" and "IID", the last two being the genotypes.}

\item{format_idx}{Column indices used for formatting haplotype string.}
}
\value{
A data frame of haplotypes
}
\description{
Infer haplotypes for a pair of SNPs
}
\author{
kaiyin
}

