% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_haplo.R
\name{getHaplos}
\alias{getHaplos}
\title{Inferring haplotypes from two genotype data frames, and join with phenotypes}
\usage{
getHaplos(g1, g2, phe, pool = NULL)
}
\arguments{
\item{g1}{First genotype data frame}

\item{g2}{Second genotype data frame, must be of the same dimension as the first. The first two column must be FID and IID.}

\item{phe}{Phenotype data frame, the first two columns must be FID and IID}

\item{pool}{A genotype data frame, assumed to be different from g1 and g2, used for pooling.}
}
\value{
A data frame containing phenotype and haplotype for each individual.
}
\description{
Inferring haplotypes from two genotype data frames, and join with phenotypes
}
\author{
kaiyin
}

