% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNshiftStr}
\alias{getNshiftStr}
\title{Extract shift number (as a string) from a shifted bed filename}
\usage{
getNshiftStr(shiftpath)
}
\arguments{
\item{shiftpath}{Path name of the shifted bed file.}
}
\value{
nshiftStr A string-representation of the shift number ("0000" for zero-shift)
}
\description{
Extract shift number (as a string) from a shifted bed filename
}

