% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{getOrElse-operator}
\alias{\%||\%}
\alias{getOrElse-operator}
\title{Default value for expression.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{expression to be evaled.}

\item{y}{default value.}
}
\description{
When an expression evals to NULL, take the default value instead. Copied from dplyr source.
}
\author{
Hadley Wickham
}

