% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_sqlUtils.R
\name{getQuery}
\alias{getQuery}
\title{Get query results from a SQLite database}
\usage{
getQuery(db_name, query_string)
}
\arguments{
\item{db_name}{character. Path to database.}

\item{query_string}{character. Query string.}
}
\description{
Get query results from a SQLite database
}
\author{
Kaiyin Zhong, Fan Liu
}

