% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{getr2}
\alias{getr2}
\title{Estimate percentage of variation explained}
\usage{
getr2(df, yn)
}
\arguments{
\item{df}{Dataframe}

\item{yn}{Name of the independent variable, must be one of the columns of \code{df}}
}
\description{
Estimate percentage of variation explained
}
\author{
Fan Liu
}

