% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getstem}
\alias{getstem}
\title{Get the stem of a bed filename.}
\usage{
getstem(pathname)
}
\arguments{
\item{pathname}{Character. Pathname of the bed file.}
}
\value{
\code{pathname} without ".bed" extension.
}
\description{
Remove the ".bed" part.
}

