% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_glm2.R
\name{glm2}
\alias{glm2}
\title{GLM with arbitrary column names}
\usage{
glm2(dat, y, xs, ...)
}
\arguments{
\item{dat}{data.frame. Souce data to build GLM upon.}

\item{y}{character. Column name of dependent variable.}

\item{xs}{character. Column names of independent variable.}

\item{...}{passed to glm.}
}
\value{
data.frame of coefficients.
}
\description{
Substitute column names that are unsuitable for formulas
and substitute back when returning results.
}
\author{
Kaiyin Zhong, Fan Liu
}

