% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_glm2.R
\name{glmIter}
\alias{glmIter}
\title{Perform glm iteratively through a number of independent variables with fixed dependent variables and covariates.}
\usage{
glmIter(dat, y, xs = NULL, covars = character(), ...)
}
\arguments{
\item{dat}{data.frame}

\item{y}{character. Name of dependent variable columns.}

\item{xs}{character. Names of independent variable columns.}

\item{covars}{character. Names of covriate columns.}

\item{...}{passed to glm.}
}
\value{
matrix of coefficients
}
\description{
Perform glm iteratively through a number of independent variables with fixed dependent variables and covariates.
}
\author{
Kaiyin Zhong, Fan Liu
}

