% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gwas}
\alias{gwas}
\title{A pure R implementation of GWAS}
\usage{
gwas(pheno_mat, geno_mat, pheno_name, covar_names)
}
\arguments{
\item{pheno_mat}{Phenotype data frame}

\item{geno_mat}{Genotype data frame}

\item{pheno_name}{Name of phenotype}

\item{covar_names}{Names of covariates}
}
\value{
A data frame with two columns, one for SNP names, one for p values.
}
\description{
A pure R implementation of GWAS
}

