% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasDat}
\alias{gwasDat}
\title{Read genotype and phenotype data into R}
\usage{
gwasDat(pl_gwas, snp_vec)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}

\item{snp_vec}{numeric or character. Vector of SNPs.}
}
\value{
data.frame
}
\description{
Read genotype and phenotype data into R
}
\author{
Kaiyin Zhong, Fan Liu
}

