% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasDir}
\alias{gwasDir}
\title{GWAS results directory of a certain GWAS scan}
\usage{
gwasDir(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object}
}
\value{
character.
}
\description{
GWAS results directory of a certain GWAS scan
}
\author{
Kaiyin Zhong, Fan Liu
}

