% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasLog}
\alias{gwasLog}
\title{Plink log fie}
\usage{
gwasLog(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}
}
\value{
character. Path to log file.
}
\description{
Redirect stdout to this file when plink is running.
}
\author{
Kaiyin Zhong, Fan Liu
}

