% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasOut}
\alias{gwasOut}
\title{GWAS output file name}
\usage{
gwasOut(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}
}
\value{
character
}
\description{
GWAS output file name
}
\author{
Kaiyin Zhong, Fan Liu
}

