% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasOutStem}
\alias{gwasOutStem}
\title{Plink output filename}
\usage{
gwasOutStem(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}
}
\value{
character. Plink output filename, without extension
}
\description{
To be passed as the \code{--out} option to plink.
}
\author{
Kaiyin Zhong, Fan Liu
}

