% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasR}
\alias{gwasR}
\title{Invoke a GWAS in R}
\usage{
gwasR(pl_gwas, snp_vec)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}

\item{snp_vec}{numeric or character. Vector of SNPs.}
}
\value{
matrix. Coefficient matrix. One row for each SNP.
}
\description{
Invoke a GWAS in R
}
\author{
Kaiyin Zhong, Fan Liu
}

