% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{gwasRDS}
\alias{gwasRDS}
\title{Get RDS file path of a PlGwasC object}
\usage{
gwasRDS(pl_gwas)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}
}
\value{
character. path of a PlGwasC object
}
\description{
Get RDS file path of a PlGwasC object
}
\author{
Kaiyin Zhong, Fan Liu
}

