% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_head2.R
\name{head2}
\alias{head2}
\alias{tail2}
\title{Head and tail in two dimensions}
\usage{
head2(x, m = 6, n = NULL)

tail2(x, m = 6, n = NULL)
}
\arguments{
\item{x}{data.frame or matrix}

\item{m}{integer. Number of rows to keep.}

\item{n}{integer. Number of columns to keep.}
}
\description{
Restrict not only the number of rows, but also the number of columns.
}
\author{
kaiyin
}

