% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{headPhe}
\alias{headPhe}
\title{Read first n lines of a phenotype file}
\usage{
headPhe(pl_gwas, nrows = 5L)
}
\arguments{
\item{pl_gwas}{PlGwasC object}

\item{nrows}{number of lines to read}
}
\value{
data.frame
}
\description{
Read first n lines of a phenotype file
}
\author{
Kaiyin Zhong, Fan Liu
}

