% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{isBinary}
\alias{isBinary}
\title{Check whether a trait is binary}
\usage{
isBinary(v, na_value = NULL)
}
\arguments{
\item{v}{numeric vector.}

\item{na_value}{a vector of numeric values which should be seen as NA.}
}
\value{
logical
}
\description{
Check whether a trait is binary
}
\examples{
\dontrun{
!isBinary(c(1, 1.1, 1, 1.1, NA))
isBinary(c(1, 2, 1, 2, NA))
!isBinary(c(-9, 2.3, 4.1, -9, -9), -9)
isBinary(c(-9, 2, 4, -9, -9), -9)
isBinary(c(1, 2, 2, 1, -9, -9.9), c(-9, -9.9))
}

}
\author{
Kaiyin Zhong, Fan Liu
}

