% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{isS4Class}
\alias{isS4Class}
\title{Check whether an S4 object is of a certain class}
\usage{
isS4Class(obj, c)
}
\arguments{
\item{obj}{S4 object}

\item{c}{Class name}
}
\value{
logical
}
\description{
Check whether an S4 object is of a certain class
}
\author{
Kaiyin Zhong, Fan Liu
}

