% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bedinfo.R
\name{isSetup}
\alias{isSetup}
\title{Check if a directory containing .bed .fam and .bim files is properly setup}
\usage{
isSetup(pl_info)
}
\arguments{
\item{pl_info}{PlInfoC object}
}
\value{
TRUE or FALSE
}
\description{
Check if a directory containing .bed .fam and .bim files is properly setup
}
\author{
Kaiyin Zhong, Fan Liu
}

