% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{isSetupRbed}
\alias{isSetupRbed}
\title{Check if an RbedInfoC object is properly set up}
\usage{
isSetupRbed(rbed_info)
}
\arguments{
\item{rbed_info}{RbedInfoC object}
}
\value{
logical.
}
\description{
Check if an RbedInfoC object is properly set up
}
\author{
Kaiyin Zhong, Fan Liu
}

