% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{isSQLite3}
\alias{isSQLite3}
\title{Check whether a file is a SQLite3 database.}
\usage{
isSQLite3(filename)
}
\arguments{
\item{filename}{character. Path to file to be checked.}
}
\description{
Check whether a file is a SQLite3 database.
}
\author{
Kaiyin Zhong, Fan Liu
}

