% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{lagDistance}
\alias{lagDistance}
\title{Distance with lag}
\usage{
lagDistance(vec, lag = 1, reverse = FALSE)
}
\arguments{
\item{vec}{numeric.}

\item{lag}{integer.}

\item{reverse}{logical. Default to FALSE, i.e. calculate \code{vec[i+lag] - vec[i]}. When set to TRUE, calculate \code{vec[i] - vec[i+lag]}}
}
\value{
numeric.
}
\description{
Calculate the distance between each element in a numeric vector and 
the element that is \code{lag} positions after it. For the last 
\code{lag} elemnts, this distance does not exist, so NA is used as a placeholder.
The returned vector is of the same length as the input vector.
}
\author{
Kaiyin Zhong, Fan Liu
}

