% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{listEqual}
\alias{listEqual}
\title{Check equality of two lists}
\usage{
listEqual(list1, list2)
}
\arguments{
\item{list1}{list}

\item{list2}{list}
}
\description{
Check equality of two lists
}
\author{
Kaiyin Zhong, Fan Liu
}

