% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{listGwasTags}
\alias{listGwasTags}
\alias{listTags}
\title{List GWAS or GCDH tags}
\usage{
listGwasTags(type = "gwas")

listTags(type = "gwas")
}
\arguments{
\item{type}{character. Either "gwas" or "gcdh".}
}
\description{
List GWAS or GCDH tags
}
\author{
Kaiyin Zhong, Fan Liu
}

