% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{loadGwas}
\alias{loadGwas}
\title{Load PlGwasC object by tag, from the RDS file}
\usage{
loadGwas(gwas_tag)
}
\arguments{
\item{gwas_tag}{character. Tag of a GWAS run.}
}
\value{
PlGwasC object.
}
\description{
Load PlGwasC object by tag, from the RDS file
}
\author{
Kaiyin Zhong, Fan Liu
}

