% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_permute.R
\name{makePhe}
\alias{makePhe}
\title{Generate phenotype file from a fam file}
\usage{
makePhe(famfile, n_components)
}
\arguments{
\item{famfile}{Character. Path of fam file.}

\item{n_components}{Integer. Number of principle components to generate.}
}
\value{
Phenotype data.frame. The data frame contains the FID, IID, SEX, AFFECTEDNESS columns of the fam file, plus principle components of genetic information.
}
\description{
Generate phenotype file from a fam file
}
\author{
kaiyin
}

