% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{manhattanData}
\alias{manhattanData}
\title{Prepare data for Manhattan plot.}
\usage{
manhattanData(chr, bp, p, snp, color_vec = NULL, sort_chr_bp = TRUE)
}
\arguments{
\item{chr}{integer. Chromosome vector.}

\item{bp}{integer. Position vector.}

\item{p}{numeric. P-value vector.}

\item{snp}{character. SNP name vector.}

\item{color_vec}{character/factor. Color vector. Doesn't have to be color names, any categorical variable will be fine.}

\item{sort_chr_bp}{logical. Whether to sort the whole data frame by CHR and BP before return.}
}
\value{
A list with the following members (1) A data frame with columns including CHR, SNP, BP, P, etc. (2) Total number of SNPs. (3) A vector of unique chromosomes.
}
\description{
Prepare data for Manhattan plot.
}
\author{
Kaiyin Zhong
}

