% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Produce Manhattan plot}
\usage{
manhattanPlot(mh_dat_res, hlines = NULL)
}
\arguments{
\item{mh_dat_res}{list. Result from \code{manhattanData}}

\item{hlines}{numeric. Horizontal lines to draw.}
}
\value{
ggplot object.
}
\description{
Produce Manhattan plot
}
\author{
Kaiyin Zhong
}

