% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bedinfo.R
\name{nIndivApprPl}
\alias{nIndivApprPl}
\title{Get apparent number of individuals}
\usage{
nIndivApprPl(pl_info)
}
\arguments{
\item{pl_info}{PlInfoC object}
}
\description{
Get apparent number of individuals
}

