% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_strConcat.R
\name{numVectorSQLRepr}
\alias{numVectorSQLRepr}
\title{String representation of a numeric vector for SQLite consumption}
\usage{
numVectorSQLRepr(vec, print_out = FALSE)
}
\arguments{
\item{vec}{numeric.}

\item{print_out}{logical. Whether to print out the string representation.}
}
\description{
Transform a numeric vector (e.g. \code{c(1, 2)} into a string representation 
that can be used in a SQLite query (e.g. "(1, 2)").
}
\author{
Kaiyin Zhong
}

