% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_permute.R
\name{permutePhe}
\alias{permutePhe}
\title{Permute a phenotype file}
\usage{
permutePhe(phe_file, out_file, force = FALSE, valid = TRUE, ...)
}
\arguments{
\item{phe_file}{character. Phenotype file.}

\item{out_file}{character. Path to permuted phenotype file.}

\item{force}{logical. When set to TRUE, existing file is overwritten.}

\item{valid}{logical. Whether to validate the phenotype file first.}

\item{...}{Passed to read.table}
}
\description{
All columns except FID and IID are permuted.
}
\author{
Kaiyin Zhong, Fan Liu
}

