% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\docType{class}
\name{PlGwasC-class}
\alias{.PlGwasC}
\alias{PlGwasC-class}
\title{An S4 class representing info about GWAS on plink files}
\description{
An S4 class representing info about GWAS on plink files
}
\section{Slots}{

\describe{
\item{\code{gwas_tag}}{character. Tag for this GWAS.}

\item{\code{opts}}{list. Plink options.}
}}

