% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bedinfo.R
\docType{class}
\name{PlInfoC-class}
\alias{.PlInfoC}
\alias{PlInfoC-class}
\title{An S4 class representing info about plink files}
\description{
Info about plink files, including the root directory, 
paths of plink .bed, .bim, .fam and .frq files, ff backing 
directories for .bim, .fam and .frq files, etc.
}
\section{Slots}{

\describe{
\item{\code{main_dir}}{Root directory where .bed, .bim and .fam files sit.}

\item{\code{plink_stem}}{character. Path to the .bed file sans the extension name}

\item{\code{plink_trio}}{character of length 3. Paths to .bed, .bim and .fam files (in that order).}

\item{\code{plink_trio_base}}{character. Basenames of \code{plink_trio}.}

\item{\code{plink_frq}}{character. Path to .frq file.}
}}

