% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_plinkr.R
\name{plinkr}
\alias{plinkr}
\title{A wrapper for plink}
\usage{
plinkr(D = NULL, K = NULL, a1_allele = NULL, a2_allele = NULL,
  adjust = NULL, all = NULL, all_pheno = NULL, allele1234 = NULL,
  alleleACGT = NULL, allele_count = NULL, allow_extra_chr = NULL,
  allow_no_sex = NULL, alt_group = NULL, alt_snp = NULL,
  annotate = NULL, annotate_snp_field = NULL, aperm = NULL,
  assoc = NULL, attrib = NULL, attrib_indiv = NULL, autosome = NULL,
  autosome_num = NULL, autosome_xy = NULL, bcf = NULL, bd = NULL,
  bed = NULL, beta = NULL, bfile = NULL, bgen = NULL,
  biallelic_only = NULL, bim = NULL, blocks = NULL,
  blocks_inform_frac = NULL, blocks_max_kb = NULL, blocks_min_maf = NULL,
  blocks_recomb_highci = NULL, blocks_strong_highci = NULL,
  blocks_strong_lowci = NULL, bmerge = NULL, border = NULL,
  bp_space = NULL, case_only = NULL, cc = NULL, cell = NULL,
  cfile = NULL, chap = NULL, check_sex = NULL, chr = NULL,
  chr_set = NULL, ci = NULL, clump = NULL, clump_allow_overlap = NULL,
  clump_annotate = NULL, clump_best = NULL, clump_field = NULL,
  clump_index_first = NULL, clump_kb = NULL, clump_p1 = NULL,
  clump_p2 = NULL, clump_r2 = NULL, clump_range = NULL,
  clump_range_border = NULL, clump_replicate = NULL,
  clump_snp_field = NULL, clump_verbose = NULL, cluster = NULL,
  cluster_missing = NULL, cm_map = NULL, cnv_blue = NULL,
  cnv_border = NULL, cnv_brown = NULL, cnv_check_no_overlap = NULL,
  cnv_count = NULL, cnv_del = NULL, cnv_disrupt = NULL,
  cnv_drop_no_segment = NULL, cnv_dup = NULL, cnv_enrichment_test = NULL,
  cnv_exclude = NULL, cnv_exclude_off_by_1 = NULL,
  cnv_freq_excldue_above = NULL, cnv_freq_excldue_below = NULL,
  cnv_freq_excldue_exact = NULL, cnv_freq_exclude_above = NULL,
  cnv_freq_exclude_below = NULL, cnv_freq_exclude_exact = NULL,
  cnv_freq_incldue_exact = NULL, cnv_freq_include_exact = NULL,
  cnv_freq_method2 = NULL, cnv_freq_overlap = NULL, cnv_green = NULL,
  cnv_indiv_perm = NULL, cnv_intersect = NULL, cnv_kb = NULL,
  cnv_list = NULL, cnv_make_map = NULL, cnv_max_kb = NULL,
  cnv_max_score = NULL, cnv_max_sites = NULL, cnv_overlap = NULL,
  cnv_red = NULL, cnv_region_overlap = NULL, cnv_report_regions = NULL,
  cnv_score = NULL, cnv_seglist = NULL, cnv_sites = NULL,
  cnv_subset = NULL, cnv_test = NULL, cnv_test_1sided = NULL,
  cnv_test_2sided = NULL, cnv_test_region = NULL, cnv_test_window = NULL,
  cnv_track = NULL, cnv_union_overlap = NULL, cnv_unique = NULL,
  cnv_verbose_report_regions = NULL, cnv_write = NULL,
  cnv_write_freq = NULL, complement_sets = NULL,
  compound_genotypes = NULL, compress = NULL, condition = NULL,
  condition_list = NULL, consensus_match = NULL, const_fid = NULL,
  control = NULL, counts = NULL, covar = NULL, covar_name = NULL,
  covar_number = NULL, cow = NULL, d = NULL, data = NULL,
  debug = NULL, decompress = NULL, dfam = NULL, distance = NULL,
  distance_exp = NULL, distance_matrix = NULL, dog = NULL,
  dominant = NULL, dosage = NULL, double_id = NULL, dprime = NULL,
  dummy = NULL, dummy_coding = NULL, each_versus_others = NULL,
  each_vs_others = NULL, epistasis = NULL, epistasis_summary_merge = NULL,
  exclude = NULL, exclude_before_extract = NULL, exclude_snp = NULL,
  exclude_snps = NULL, extract = NULL, fam = NULL, family = NULL,
  fast_epistasis = NULL, fid = NULL, file = NULL,
  fill_missing_a2 = NULL, filter = NULL, filter_cases = NULL,
  filter_controls = NULL, filter_females = NULL, filter_founders = NULL,
  filter_males = NULL, filter_nonfounders = NULL, fisher = NULL,
  flip = NULL, flip_scan = NULL, flip_scan_threshold = NULL,
  flip_scan_verbose = NULL, flip_scan_window = NULL,
  flip_scan_window_kb = NULL, flip_subset = NULL, freq = NULL,
  freqx = NULL, from = NULL, from_bp = NULL, from_kb = NULL,
  from_mb = NULL, frqx = NULL, fst = NULL, gap = NULL, gates = NULL,
  gc = NULL, gen = NULL, gene = NULL, gene_all = NULL,
  gene_list = NULL, gene_list_border = NULL, gene_report = NULL,
  gene_report_empty = NULL, gene_report_snp_field = NULL,
  gene_subset = NULL, genedrop = NULL, genepi = NULL, geno = NULL,
  genome = NULL, genome_full = NULL, genome_lists = NULL,
  genome_minimal = NULL, genotypic = NULL, gfile = NULL, gplink = NULL,
  grm = NULL, grm_bin = NULL, grm_gz = NULL, group_avg = NULL,
  groupdist = NULL, gxe = NULL, hap... = NULL, hap = NULL,
  hap_assoc = NULL, hap_freq = NULL, hap_impute = NULL,
  hap_max_phase = NULL, hap_min_phase_prob = NULL, hap_miss = NULL,
  hap_phase = NULL, hap_phase_wide = NULL, hap_pp = NULL,
  hap_snps = NULL, hap_tdt = NULL, hap_window = NULL,
  hard_call_threshold = NULL, hardy2 = NULL, hardy = NULL, help = NULL,
  het = NULL, hethom = NULL, hide_covar = NULL, homog = NULL,
  homozyg = NULL, homozyg_density = NULL, homozyg_gap = NULL,
  homozyg_group = NULL, homozyg_het = NULL,
  homozyg_include_missing = NULL, homozyg_kb = NULL, homozyg_match = NULL,
  homozyg_snp = NULL, homozyg_verbose = NULL, homozyg_window_het = NULL,
  homozyg_window_kb = NULL, homozyg_window_missing = NULL,
  homozyg_window_snp = NULL, homozyg_window_threshold = NULL,
  horse = NULL, hwe = NULL, hwe_all = NULL, ibc = NULL, ibm = NULL,
  ibs_matrix = NULL, ibs_test = NULL, id_delim = NULL, id_dict = NULL,
  id_match = NULL, iid = NULL, impossible = NULL, impute_sex = NULL,
  ind_major = NULL, indep = NULL, indep_pairphase = NULL,
  indep_pairwise = NULL, independent_effect = NULL, indiv_sort = NULL,
  inter_chr = NULL, interaction = NULL, je_cellmin = NULL, keep = NULL,
  keep_allele_order = NULL, keep_autoconv = NULL,
  keep_before_remove = NULL, keep_cluster_names = NULL,
  keep_clusters = NULL, keep_fam = NULL, lambda = NULL, lasso = NULL,
  lasso_select_covars = NULL, ld = NULL, ld_snp = NULL,
  ld_snp_list = NULL, ld_snps = NULL, ld_window = NULL,
  ld_window_kb = NULL, ld_window_r2 = NULL, ld_xchr = NULL,
  lfile = NULL, liability = NULL, linear = NULL, list = NULL,
  list_23_indels = NULL, list_all = NULL, logistic = NULL,
  lookup... = NULL, lookup = NULL, lookup_gene = NULL,
  lookup_list = NULL, loop_assoc = NULL, maf = NULL, maf_succ = NULL,
  make_bed = NULL, make_founders = NULL, make_grm = NULL,
  make_grm_bin = NULL, make_grm_gz = NULL, make_just_bim = NULL,
  make_just_fam = NULL, make_perm_pheno = NULL, make_pheno = NULL,
  make_rel = NULL, make_set = NULL, make_set_border = NULL,
  make_set_collapse_group = NULL, make_set_complement_all = NULL,
  make_set_complement_group = NULL, map = NULL, mat = NULL,
  match = NULL, match_type = NULL, matrix = NULL, max = NULL,
  max_maf = NULL, mc = NULL, mcc = NULL, mcovar = NULL,
  mds_cluster = NULL, mds_plot = NULL, me = NULL, me_exclude_one = NULL,
  memory = NULL, mendel = NULL, mendel_duos = NULL,
  mendel_multigen = NULL, merge = NULL, merge_equal_pos = NULL,
  merge_list = NULL, merge_mode = NULL, merge_x = NULL,
  meta_analysis = NULL, meta_analysis_..._field = NULL, mfilter = NULL,
  mh = NULL, mhf = NULL, min = NULL, mind = NULL,
  mishap_window = NULL, missing = NULL, missing_code = NULL,
  missing_genotype = NULL, missing_phenotype = NULL,
  missing_var_code = NULL, mlma = NULL, mlma_loco = NULL,
  mlma_no_adj_covar = NULL, model = NULL, model_dom = NULL,
  model_gen = NULL, model_rec = NULL, model_trend = NULL, mouse = NULL,
  mperm = NULL, mperm_save = NULL, mperm_save_all = NULL, mpheno = NULL,
  must_have_sex = NULL, mwithin = NULL, neighbour = NULL, no_fid = NULL,
  no_parents = NULL, no_pheno = NULL, no_sex = NULL, no_snp = NULL,
  no_x_sex = NULL, nonfounders = NULL, nop = NULL, not_chr = NULL,
  nudge = NULL, null_group = NULL, null_snp = NULL,
  oblig_cluster = NULL, oblig_clusters = NULL, oblig_missing = NULL,
  out = NULL, output_chr = NULL, output_missing_genotype = NULL,
  output_missing_phenotype = NULL, oxford_pheno_name = NULL,
  parallel = NULL, parameters = NULL, parentdt1 = NULL,
  parentdt2 = NULL, pat = NULL, pca = NULL, pca_cluster_names = NULL,
  pca_clusters = NULL, ped = NULL, pedigree = NULL, perm = NULL,
  perm_batch_size = NULL, perm_count = NULL, pfilter = NULL,
  pheno = NULL, pheno_merge = NULL, pheno_name = NULL, pick1 = NULL,
  plist = NULL, poo = NULL, pool_size = NULL, ppc = NULL,
  ppc_gap = NULL, proxy_... = NULL, proxy_assoc = NULL,
  proxy_b_kb = NULL, proxy_b_maxsnp = NULL, proxy_b_r2 = NULL,
  proxy_b_threshold = NULL, proxy_b_window = NULL, proxy_dosage = NULL,
  proxy_drop = NULL, proxy_flanking = NULL, proxy_geno = NULL,
  proxy_genotypic_concordance = NULL, proxy_glm = NULL,
  proxy_impute = NULL, proxy_impute_threshold = NULL, proxy_kb = NULL,
  proxy_list = NULL, proxy_maf = NULL, proxy_maxsnp = NULL,
  proxy_mhf = NULL, proxy_r2 = NULL, proxy_r2_no_filter = NULL,
  proxy_replace = NULL, proxy_show_proxies = NULL,
  proxy_sub_maxsnp = NULL, proxy_sub_r2 = NULL, proxy_tdt = NULL,
  proxy_verbose = NULL, proxy_window = NULL, prune = NULL,
  q_score_file = NULL, q_score_range = NULL, qfam... = NULL,
  qmatch = NULL, qq_plot = NULL, qt = NULL, qt_means = NULL,
  qual_geno_... = NULL, qual_geno_max_threshold = NULL,
  qual_geno_scores = NULL, qual_geno_threshold = NULL,
  qual_max_threshold = NULL, qual_scores = NULL, qual_threshold = NULL,
  r2 = NULL, r = NULL, range = NULL, rank = NULL, read_dists = NULL,
  read_freq = NULL, read_genome = NULL, read_genome_list = NULL,
  read_genome_minimal = NULL, recessive = NULL, recode12 = NULL,
  recode = NULL, recodeA = NULL, recodeAD = NULL, recodeHV = NULL,
  recode_allele = NULL, recode_beagle = NULL, recode_bimbam = NULL,
  recode_fastphase = NULL, recode_lgen = NULL, recode_rlist = NULL,
  recode_structure = NULL, recode_vcf = NULL, recode_whap = NULL,
  reference = NULL, reference_allele = NULL, regress_distance = NULL,
  regress_pcs = NULL, regress_rel = NULL, rel_check = NULL,
  rel_cutoff = NULL, remove = NULL, remove_cluster_names = NULL,
  remove_clusters = NULL, remove_fam = NULL, rerun = NULL, rice = NULL,
  sample = NULL, score = NULL, score_no_mean_imputation = NULL,
  script = NULL, seed = NULL, set = NULL, set_by_all = NULL,
  set_collapse_all = NULL, set_hh_missing = NULL, set_max = NULL,
  set_me_missing = NULL, set_missing_nonsnp_ids = NULL,
  set_missing_snp_ids = NULL, set_missing_var_ids = NULL,
  set_names = NULL, set_p = NULL, set_r2 = NULL, set_r2_phase = NULL,
  set_table = NULL, set_test = NULL, sex = NULL, sheep = NULL,
  show_tags = NULL, silent = NULL, simulate = NULL,
  simulate_haps = NULL, simulate_label = NULL, simulate_missing = NULL,
  simulate_n = NULL, simulate_ncases = NULL, simulate_ncontrols = NULL,
  simulate_prevalence = NULL, simulate_qt = NULL, simulate_tags = NULL,
  snp = NULL, snps = NULL, snps_only = NULL, specific_haplotype = NULL,
  split_x = NULL, standard_beta = NULL, subset = NULL,
  swap_parents = NULL, swap_sibs = NULL, swap_unrel = NULL, tab = NULL,
  tag_kb = NULL, tag_mode2 = NULL, tag_r2 = NULL, tail_pheno = NULL,
  tdt = NULL, test_all = NULL, test_mishap = NULL, test_missing = NULL,
  test_snp = NULL, tests = NULL, tfam = NULL, tfile = NULL,
  thin = NULL, thin_count = NULL, threads = NULL, to = NULL,
  to_bp = NULL, to_kb = NULL, to_mb = NULL, tped = NULL,
  transpose = NULL, trend = NULL, tucc = NULL, twolocus = NULL,
  unbounded = NULL, unrelated_heritability = NULL, update_alleles = NULL,
  update_chr = NULL, update_cm = NULL, update_ids = NULL,
  update_map = NULL, update_name = NULL, update_parents = NULL,
  update_sex = NULL, vcf = NULL, vcf_filter = NULL,
  vcf_half_call = NULL, vcf_idspace_to = NULL, vcf_min_qual = NULL,
  vegas = NULL, version = NULL, vif = NULL, whap = NULL,
  window = NULL, with_freqs = NULL, with_phenotype = NULL,
  with_reference = NULL, within = NULL, write_cluster = NULL,
  write_covar = NULL, write_dosage = NULL, write_set = NULL,
  write_set_r2 = NULL, write_snplist = NULL, xchr_model = NULL,
  zero_cluster = NULL, zero_cms = NULL, one = NULL, twothreefile = NULL,
  stdout = collenv$.plink_stdout, stderr = collenv$.plink_stderr,
  wait = TRUE)
}
\arguments{
\item{D}{Same as plink --D}

\item{K}{Same as plink --K}

\item{a1_allele}{Same as plink --a1-allele}

\item{a2_allele}{Same as plink --a2-allele}

\item{adjust}{Same as plink --adjust}

\item{all}{Same as plink --all}

\item{all_pheno}{Same as plink --all-pheno}

\item{allele1234}{Same as plink --allele1234}

\item{alleleACGT}{Same as plink --alleleACGT}

\item{allele_count}{Same as plink --allele-count}

\item{allow_extra_chr}{Same as plink --allow-extra-chr}

\item{allow_no_sex}{Same as plink --allow-no-sex}

\item{alt_group}{Same as plink --alt-group}

\item{alt_snp}{Same as plink --alt-snp}

\item{annotate}{Same as plink --annotate}

\item{annotate_snp_field}{Same as plink --annotate-snp-field}

\item{aperm}{Same as plink --aperm}

\item{assoc}{Same as plink --assoc}

\item{attrib}{Same as plink --attrib}

\item{attrib_indiv}{Same as plink --attrib-indiv}

\item{autosome}{Same as plink --autosome}

\item{autosome_num}{Same as plink --autosome-num}

\item{autosome_xy}{Same as plink --autosome-xy}

\item{bcf}{Same as plink --bcf}

\item{bd}{Same as plink --bd}

\item{bed}{Same as plink --bed}

\item{beta}{Same as plink --beta}

\item{bfile}{Same as plink --bfile}

\item{bgen}{Same as plink --bgen}

\item{biallelic_only}{Same as plink --biallelic-only}

\item{bim}{Same as plink --bim}

\item{blocks}{Same as plink --blocks}

\item{blocks_inform_frac}{Same as plink --blocks-inform-frac}

\item{blocks_max_kb}{Same as plink --blocks-max-kb}

\item{blocks_min_maf}{Same as plink --blocks-min-maf}

\item{blocks_recomb_highci}{Same as plink --blocks-recomb-highci}

\item{blocks_strong_highci}{Same as plink --blocks-strong-highci}

\item{blocks_strong_lowci}{Same as plink --blocks-strong-lowci}

\item{bmerge}{Same as plink --bmerge}

\item{border}{Same as plink --border}

\item{bp_space}{Same as plink --bp-space}

\item{case_only}{Same as plink --case-only}

\item{cc}{Same as plink --cc}

\item{cell}{Same as plink --cell}

\item{cfile}{Same as plink --cfile}

\item{chap}{Same as plink --chap}

\item{check_sex}{Same as plink --check-sex}

\item{chr}{Same as plink --chr}

\item{chr_set}{Same as plink --chr-set}

\item{ci}{Same as plink --ci}

\item{clump}{Same as plink --clump}

\item{clump_allow_overlap}{Same as plink --clump-allow-overlap}

\item{clump_annotate}{Same as plink --clump-annotate}

\item{clump_best}{Same as plink --clump-best}

\item{clump_field}{Same as plink --clump-field}

\item{clump_index_first}{Same as plink --clump-index-first}

\item{clump_kb}{Same as plink --clump-kb}

\item{clump_p1}{Same as plink --clump-p1}

\item{clump_p2}{Same as plink --clump-p2}

\item{clump_r2}{Same as plink --clump-r2}

\item{clump_range}{Same as plink --clump-range}

\item{clump_range_border}{Same as plink --clump-range-border}

\item{clump_replicate}{Same as plink --clump-replicate}

\item{clump_snp_field}{Same as plink --clump-snp-field}

\item{clump_verbose}{Same as plink --clump-verbose}

\item{cluster}{Same as plink --cluster}

\item{cluster_missing}{Same as plink --cluster-missing}

\item{cm_map}{Same as plink --cm-map}

\item{cnv_blue}{Same as plink --cnv-blue}

\item{cnv_border}{Same as plink --cnv-border}

\item{cnv_brown}{Same as plink --cnv-brown}

\item{cnv_check_no_overlap}{Same as plink --cnv-check-no-overlap}

\item{cnv_count}{Same as plink --cnv-count}

\item{cnv_del}{Same as plink --cnv-del}

\item{cnv_disrupt}{Same as plink --cnv-disrupt}

\item{cnv_drop_no_segment}{Same as plink --cnv-drop-no-segment}

\item{cnv_dup}{Same as plink --cnv-dup}

\item{cnv_enrichment_test}{Same as plink --cnv-enrichment-test}

\item{cnv_exclude}{Same as plink --cnv-exclude}

\item{cnv_exclude_off_by_1}{Same as plink --cnv-exclude-off-by-1}

\item{cnv_freq_excldue_above}{Same as plink --cnv-freq-excldue-above}

\item{cnv_freq_excldue_below}{Same as plink --cnv-freq-excldue-below}

\item{cnv_freq_excldue_exact}{Same as plink --cnv-freq-excldue-exact}

\item{cnv_freq_exclude_above}{Same as plink --cnv-freq-exclude-above}

\item{cnv_freq_exclude_below}{Same as plink --cnv-freq-exclude-below}

\item{cnv_freq_exclude_exact}{Same as plink --cnv-freq-exclude-exact}

\item{cnv_freq_incldue_exact}{Same as plink --cnv-freq-incldue-exact}

\item{cnv_freq_include_exact}{Same as plink --cnv-freq-include-exact}

\item{cnv_freq_method2}{Same as plink --cnv-freq-method2}

\item{cnv_freq_overlap}{Same as plink --cnv-freq-overlap}

\item{cnv_green}{Same as plink --cnv-green}

\item{cnv_indiv_perm}{Same as plink --cnv-indiv-perm}

\item{cnv_intersect}{Same as plink --cnv-intersect}

\item{cnv_kb}{Same as plink --cnv-kb}

\item{cnv_list}{Same as plink --cnv-list}

\item{cnv_make_map}{Same as plink --cnv-make-map}

\item{cnv_max_kb}{Same as plink --cnv-max-kb}

\item{cnv_max_score}{Same as plink --cnv-max-score}

\item{cnv_max_sites}{Same as plink --cnv-max-sites}

\item{cnv_overlap}{Same as plink --cnv-overlap}

\item{cnv_red}{Same as plink --cnv-red}

\item{cnv_region_overlap}{Same as plink --cnv-region-overlap}

\item{cnv_report_regions}{Same as plink --cnv-report-regions}

\item{cnv_score}{Same as plink --cnv-score}

\item{cnv_seglist}{Same as plink --cnv-seglist}

\item{cnv_sites}{Same as plink --cnv-sites}

\item{cnv_subset}{Same as plink --cnv-subset}

\item{cnv_test}{Same as plink --cnv-test}

\item{cnv_test_1sided}{Same as plink --cnv-test-1sided}

\item{cnv_test_2sided}{Same as plink --cnv-test-2sided}

\item{cnv_test_region}{Same as plink --cnv-test-region}

\item{cnv_test_window}{Same as plink --cnv-test-window}

\item{cnv_track}{Same as plink --cnv-track}

\item{cnv_union_overlap}{Same as plink --cnv-union-overlap}

\item{cnv_unique}{Same as plink --cnv-unique}

\item{cnv_verbose_report_regions}{Same as plink --cnv-verbose-report-regions}

\item{cnv_write}{Same as plink --cnv-write}

\item{cnv_write_freq}{Same as plink --cnv-write-freq}

\item{complement_sets}{Same as plink --complement-sets}

\item{compound_genotypes}{Same as plink --compound-genotypes}

\item{compress}{Same as plink --compress}

\item{condition}{Same as plink --condition}

\item{condition_list}{Same as plink --condition-list}

\item{consensus_match}{Same as plink --consensus-match}

\item{const_fid}{Same as plink --const-fid}

\item{control}{Same as plink --control}

\item{counts}{Same as plink --counts}

\item{covar}{Same as plink --covar}

\item{covar_name}{Same as plink --covar-name}

\item{covar_number}{Same as plink --covar-number}

\item{cow}{Same as plink --cow}

\item{d}{Same as plink --d}

\item{data}{Same as plink --data}

\item{debug}{Same as plink --debug}

\item{decompress}{Same as plink --decompress}

\item{dfam}{Same as plink --dfam}

\item{distance}{Same as plink --distance}

\item{distance_exp}{Same as plink --distance-exp}

\item{distance_matrix}{Same as plink --distance-matrix}

\item{dog}{Same as plink --dog}

\item{dominant}{Same as plink --dominant}

\item{dosage}{Same as plink --dosage}

\item{double_id}{Same as plink --double-id}

\item{dprime}{Same as plink --dprime}

\item{dummy}{Same as plink --dummy}

\item{dummy_coding}{Same as plink --dummy-coding}

\item{each_versus_others}{Same as plink --each-versus-others}

\item{each_vs_others}{Same as plink --each-vs-others}

\item{epistasis}{Same as plink --epistasis}

\item{epistasis_summary_merge}{Same as plink --epistasis-summary-merge}

\item{exclude}{Same as plink --exclude}

\item{exclude_before_extract}{Same as plink --exclude-before-extract}

\item{exclude_snp}{Same as plink --exclude-snp}

\item{exclude_snps}{Same as plink --exclude-snps}

\item{extract}{Same as plink --extract}

\item{fam}{Same as plink --fam}

\item{family}{Same as plink --family}

\item{fast_epistasis}{Same as plink --fast-epistasis}

\item{fid}{Same as plink --fid}

\item{file}{Same as plink --file}

\item{fill_missing_a2}{Same as plink --fill-missing-a2}

\item{filter}{Same as plink --filter}

\item{filter_cases}{Same as plink --filter-cases}

\item{filter_controls}{Same as plink --filter-controls}

\item{filter_females}{Same as plink --filter-females}

\item{filter_founders}{Same as plink --filter-founders}

\item{filter_males}{Same as plink --filter-males}

\item{filter_nonfounders}{Same as plink --filter-nonfounders}

\item{fisher}{Same as plink --fisher}

\item{flip}{Same as plink --flip}

\item{flip_scan}{Same as plink --flip-scan}

\item{flip_scan_threshold}{Same as plink --flip-scan-threshold}

\item{flip_scan_verbose}{Same as plink --flip-scan-verbose}

\item{flip_scan_window}{Same as plink --flip-scan-window}

\item{flip_scan_window_kb}{Same as plink --flip-scan-window-kb}

\item{flip_subset}{Same as plink --flip-subset}

\item{freq}{Same as plink --freq}

\item{freqx}{Same as plink --freqx}

\item{from}{Same as plink --from}

\item{from_bp}{Same as plink --from-bp}

\item{from_kb}{Same as plink --from-kb}

\item{from_mb}{Same as plink --from-mb}

\item{frqx}{Same as plink --frqx}

\item{fst}{Same as plink --fst}

\item{gap}{Same as plink --gap}

\item{gates}{Same as plink --gates}

\item{gc}{Same as plink --gc}

\item{gen}{Same as plink --gen}

\item{gene}{Same as plink --gene}

\item{gene_all}{Same as plink --gene-all}

\item{gene_list}{Same as plink --gene-list}

\item{gene_list_border}{Same as plink --gene-list-border}

\item{gene_report}{Same as plink --gene-report}

\item{gene_report_empty}{Same as plink --gene-report-empty}

\item{gene_report_snp_field}{Same as plink --gene-report-snp-field}

\item{gene_subset}{Same as plink --gene-subset}

\item{genedrop}{Same as plink --genedrop}

\item{genepi}{Same as plink --genepi}

\item{geno}{Same as plink --geno}

\item{genome}{Same as plink --genome}

\item{genome_full}{Same as plink --genome-full}

\item{genome_lists}{Same as plink --genome-lists}

\item{genome_minimal}{Same as plink --genome-minimal}

\item{genotypic}{Same as plink --genotypic}

\item{gfile}{Same as plink --gfile}

\item{gplink}{Same as plink --gplink}

\item{grm}{Same as plink --grm}

\item{grm_bin}{Same as plink --grm-bin}

\item{grm_gz}{Same as plink --grm-gz}

\item{group_avg}{Same as plink --group-avg}

\item{groupdist}{Same as plink --groupdist}

\item{gxe}{Same as plink --gxe}

\item{hap...}{Same as plink --hap...}

\item{hap}{Same as plink --hap}

\item{hap_assoc}{Same as plink --hap-assoc}

\item{hap_freq}{Same as plink --hap-freq}

\item{hap_impute}{Same as plink --hap-impute}

\item{hap_max_phase}{Same as plink --hap-max-phase}

\item{hap_min_phase_prob}{Same as plink --hap-min-phase-prob}

\item{hap_miss}{Same as plink --hap-miss}

\item{hap_phase}{Same as plink --hap-phase}

\item{hap_phase_wide}{Same as plink --hap-phase-wide}

\item{hap_pp}{Same as plink --hap-pp}

\item{hap_snps}{Same as plink --hap-snps}

\item{hap_tdt}{Same as plink --hap-tdt}

\item{hap_window}{Same as plink --hap-window}

\item{hard_call_threshold}{Same as plink --hard-call-threshold}

\item{hardy2}{Same as plink --hardy2}

\item{hardy}{Same as plink --hardy}

\item{help}{Same as plink --help}

\item{het}{Same as plink --het}

\item{hethom}{Same as plink --hethom}

\item{hide_covar}{Same as plink --hide-covar}

\item{homog}{Same as plink --homog}

\item{homozyg}{Same as plink --homozyg}

\item{homozyg_density}{Same as plink --homozyg-density}

\item{homozyg_gap}{Same as plink --homozyg-gap}

\item{homozyg_group}{Same as plink --homozyg-group}

\item{homozyg_het}{Same as plink --homozyg-het}

\item{homozyg_include_missing}{Same as plink --homozyg-include-missing}

\item{homozyg_kb}{Same as plink --homozyg-kb}

\item{homozyg_match}{Same as plink --homozyg-match}

\item{homozyg_snp}{Same as plink --homozyg-snp}

\item{homozyg_verbose}{Same as plink --homozyg-verbose}

\item{homozyg_window_het}{Same as plink --homozyg-window-het}

\item{homozyg_window_kb}{Same as plink --homozyg-window-kb}

\item{homozyg_window_missing}{Same as plink --homozyg-window-missing}

\item{homozyg_window_snp}{Same as plink --homozyg-window-snp}

\item{homozyg_window_threshold}{Same as plink --homozyg-window-threshold}

\item{horse}{Same as plink --horse}

\item{hwe}{Same as plink --hwe}

\item{hwe_all}{Same as plink --hwe-all}

\item{ibc}{Same as plink --ibc}

\item{ibm}{Same as plink --ibm}

\item{ibs_matrix}{Same as plink --ibs-matrix}

\item{ibs_test}{Same as plink --ibs-test}

\item{id_delim}{Same as plink --id-delim}

\item{id_dict}{Same as plink --id-dict}

\item{id_match}{Same as plink --id-match}

\item{iid}{Same as plink --iid}

\item{impossible}{Same as plink --impossible}

\item{impute_sex}{Same as plink --impute-sex}

\item{ind_major}{Same as plink --ind-major}

\item{indep}{Same as plink --indep}

\item{indep_pairphase}{Same as plink --indep-pairphase}

\item{indep_pairwise}{Same as plink --indep-pairwise}

\item{independent_effect}{Same as plink --independent-effect}

\item{indiv_sort}{Same as plink --indiv-sort}

\item{inter_chr}{Same as plink --inter-chr}

\item{interaction}{Same as plink --interaction}

\item{je_cellmin}{Same as plink --je-cellmin}

\item{keep}{Same as plink --keep}

\item{keep_allele_order}{Same as plink --keep-allele-order}

\item{keep_autoconv}{Same as plink --keep-autoconv}

\item{keep_before_remove}{Same as plink --keep-before-remove}

\item{keep_cluster_names}{Same as plink --keep-cluster-names}

\item{keep_clusters}{Same as plink --keep-clusters}

\item{keep_fam}{Same as plink --keep-fam}

\item{lambda}{Same as plink --lambda}

\item{lasso}{Same as plink --lasso}

\item{lasso_select_covars}{Same as plink --lasso-select-covars}

\item{ld}{Same as plink --ld}

\item{ld_snp}{Same as plink --ld-snp}

\item{ld_snp_list}{Same as plink --ld-snp-list}

\item{ld_snps}{Same as plink --ld-snps}

\item{ld_window}{Same as plink --ld-window}

\item{ld_window_kb}{Same as plink --ld-window-kb}

\item{ld_window_r2}{Same as plink --ld-window-r2}

\item{ld_xchr}{Same as plink --ld-xchr}

\item{lfile}{Same as plink --lfile}

\item{liability}{Same as plink --liability}

\item{linear}{Same as plink --linear}

\item{list}{Same as plink --list}

\item{list_23_indels}{Same as plink --list-23-indels}

\item{list_all}{Same as plink --list-all}

\item{logistic}{Same as plink --logistic}

\item{lookup...}{Same as plink --lookup...}

\item{lookup}{Same as plink --lookup}

\item{lookup_gene}{Same as plink --lookup-gene}

\item{lookup_list}{Same as plink --lookup-list}

\item{loop_assoc}{Same as plink --loop-assoc}

\item{maf}{Same as plink --maf}

\item{maf_succ}{Same as plink --maf-succ}

\item{make_bed}{Same as plink --make-bed}

\item{make_founders}{Same as plink --make-founders}

\item{make_grm}{Same as plink --make-grm}

\item{make_grm_bin}{Same as plink --make-grm-bin}

\item{make_grm_gz}{Same as plink --make-grm-gz}

\item{make_just_bim}{Same as plink --make-just-bim}

\item{make_just_fam}{Same as plink --make-just-fam}

\item{make_perm_pheno}{Same as plink --make-perm-pheno}

\item{make_pheno}{Same as plink --make-pheno}

\item{make_rel}{Same as plink --make-rel}

\item{make_set}{Same as plink --make-set}

\item{make_set_border}{Same as plink --make-set-border}

\item{make_set_collapse_group}{Same as plink --make-set-collapse-group}

\item{make_set_complement_all}{Same as plink --make-set-complement-all}

\item{make_set_complement_group}{Same as plink --make-set-complement-group}

\item{map}{Same as plink --map}

\item{mat}{Same as plink --mat}

\item{match}{Same as plink --match}

\item{match_type}{Same as plink --match-type}

\item{matrix}{Same as plink --matrix}

\item{max}{Same as plink --max}

\item{max_maf}{Same as plink --max-maf}

\item{mc}{Same as plink --mc}

\item{mcc}{Same as plink --mcc}

\item{mcovar}{Same as plink --mcovar}

\item{mds_cluster}{Same as plink --mds-cluster}

\item{mds_plot}{Same as plink --mds-plot}

\item{me}{Same as plink --me}

\item{me_exclude_one}{Same as plink --me-exclude-one}

\item{memory}{Same as plink --memory}

\item{mendel}{Same as plink --mendel}

\item{mendel_duos}{Same as plink --mendel-duos}

\item{mendel_multigen}{Same as plink --mendel-multigen}

\item{merge}{Same as plink --merge}

\item{merge_equal_pos}{Same as plink --merge-equal-pos}

\item{merge_list}{Same as plink --merge-list}

\item{merge_mode}{Same as plink --merge-mode}

\item{merge_x}{Same as plink --merge-x}

\item{meta_analysis}{Same as plink --meta-analysis}

\item{meta_analysis_..._field}{Same as plink --meta-analysis-...-field}

\item{mfilter}{Same as plink --mfilter}

\item{mh}{Same as plink --mh}

\item{mhf}{Same as plink --mhf}

\item{min}{Same as plink --min}

\item{mind}{Same as plink --mind}

\item{mishap_window}{Same as plink --mishap-window}

\item{missing}{Same as plink --missing}

\item{missing_code}{Same as plink --missing-code}

\item{missing_genotype}{Same as plink --missing-genotype}

\item{missing_phenotype}{Same as plink --missing-phenotype}

\item{missing_var_code}{Same as plink --missing-var-code}

\item{mlma}{Same as plink --mlma}

\item{mlma_loco}{Same as plink --mlma-loco}

\item{mlma_no_adj_covar}{Same as plink --mlma-no-adj-covar}

\item{model}{Same as plink --model}

\item{model_dom}{Same as plink --model-dom}

\item{model_gen}{Same as plink --model-gen}

\item{model_rec}{Same as plink --model-rec}

\item{model_trend}{Same as plink --model-trend}

\item{mouse}{Same as plink --mouse}

\item{mperm}{Same as plink --mperm}

\item{mperm_save}{Same as plink --mperm-save}

\item{mperm_save_all}{Same as plink --mperm-save-all}

\item{mpheno}{Same as plink --mpheno}

\item{must_have_sex}{Same as plink --must-have-sex}

\item{mwithin}{Same as plink --mwithin}

\item{neighbour}{Same as plink --neighbour}

\item{no_fid}{Same as plink --no-fid}

\item{no_parents}{Same as plink --no-parents}

\item{no_pheno}{Same as plink --no-pheno}

\item{no_sex}{Same as plink --no-sex}

\item{no_snp}{Same as plink --no-snp}

\item{no_x_sex}{Same as plink --no-x-sex}

\item{nonfounders}{Same as plink --nonfounders}

\item{nop}{Same as plink --nop}

\item{not_chr}{Same as plink --not-chr}

\item{nudge}{Same as plink --nudge}

\item{null_group}{Same as plink --null-group}

\item{null_snp}{Same as plink --null-snp}

\item{oblig_cluster}{Same as plink --oblig-cluster}

\item{oblig_clusters}{Same as plink --oblig-clusters}

\item{oblig_missing}{Same as plink --oblig-missing}

\item{out}{Same as plink --out}

\item{output_chr}{Same as plink --output-chr}

\item{output_missing_genotype}{Same as plink --output-missing-genotype}

\item{output_missing_phenotype}{Same as plink --output-missing-phenotype}

\item{oxford_pheno_name}{Same as plink --oxford-pheno-name}

\item{parallel}{Same as plink --parallel}

\item{parameters}{Same as plink --parameters}

\item{parentdt1}{Same as plink --parentdt1}

\item{parentdt2}{Same as plink --parentdt2}

\item{pat}{Same as plink --pat}

\item{pca}{Same as plink --pca}

\item{pca_cluster_names}{Same as plink --pca-cluster-names}

\item{pca_clusters}{Same as plink --pca-clusters}

\item{ped}{Same as plink --ped}

\item{pedigree}{Same as plink --pedigree}

\item{perm}{Same as plink --perm}

\item{perm_batch_size}{Same as plink --perm-batch-size}

\item{perm_count}{Same as plink --perm-count}

\item{pfilter}{Same as plink --pfilter}

\item{pheno}{Same as plink --pheno}

\item{pheno_merge}{Same as plink --pheno-merge}

\item{pheno_name}{Same as plink --pheno-name}

\item{pick1}{Same as plink --pick1}

\item{plist}{Same as plink --plist}

\item{poo}{Same as plink --poo}

\item{pool_size}{Same as plink --pool-size}

\item{ppc}{Same as plink --ppc}

\item{ppc_gap}{Same as plink --ppc-gap}

\item{proxy_...}{Same as plink --proxy-...}

\item{proxy_assoc}{Same as plink --proxy-assoc}

\item{proxy_b_kb}{Same as plink --proxy-b-kb}

\item{proxy_b_maxsnp}{Same as plink --proxy-b-maxsnp}

\item{proxy_b_r2}{Same as plink --proxy-b-r2}

\item{proxy_b_threshold}{Same as plink --proxy-b-threshold}

\item{proxy_b_window}{Same as plink --proxy-b-window}

\item{proxy_dosage}{Same as plink --proxy-dosage}

\item{proxy_drop}{Same as plink --proxy-drop}

\item{proxy_flanking}{Same as plink --proxy-flanking}

\item{proxy_geno}{Same as plink --proxy-geno}

\item{proxy_genotypic_concordance}{Same as plink --proxy-genotypic-concordance}

\item{proxy_glm}{Same as plink --proxy-glm}

\item{proxy_impute}{Same as plink --proxy-impute}

\item{proxy_impute_threshold}{Same as plink --proxy-impute-threshold}

\item{proxy_kb}{Same as plink --proxy-kb}

\item{proxy_list}{Same as plink --proxy-list}

\item{proxy_maf}{Same as plink --proxy-maf}

\item{proxy_maxsnp}{Same as plink --proxy-maxsnp}

\item{proxy_mhf}{Same as plink --proxy-mhf}

\item{proxy_r2}{Same as plink --proxy-r2}

\item{proxy_r2_no_filter}{Same as plink --proxy-r2-no-filter}

\item{proxy_replace}{Same as plink --proxy-replace}

\item{proxy_show_proxies}{Same as plink --proxy-show-proxies}

\item{proxy_sub_maxsnp}{Same as plink --proxy-sub-maxsnp}

\item{proxy_sub_r2}{Same as plink --proxy-sub-r2}

\item{proxy_tdt}{Same as plink --proxy-tdt}

\item{proxy_verbose}{Same as plink --proxy-verbose}

\item{proxy_window}{Same as plink --proxy-window}

\item{prune}{Same as plink --prune}

\item{q_score_file}{Same as plink --q-score-file}

\item{q_score_range}{Same as plink --q-score-range}

\item{qfam...}{Same as plink --qfam...}

\item{qmatch}{Same as plink --qmatch}

\item{qq_plot}{Same as plink --qq-plot}

\item{qt}{Same as plink --qt}

\item{qt_means}{Same as plink --qt-means}

\item{qual_geno_...}{Same as plink --qual-geno-...}

\item{qual_geno_max_threshold}{Same as plink --qual-geno-max-threshold}

\item{qual_geno_scores}{Same as plink --qual-geno-scores}

\item{qual_geno_threshold}{Same as plink --qual-geno-threshold}

\item{qual_max_threshold}{Same as plink --qual-max-threshold}

\item{qual_scores}{Same as plink --qual-scores}

\item{qual_threshold}{Same as plink --qual-threshold}

\item{r2}{Same as plink --r2}

\item{r}{Same as plink --r}

\item{range}{Same as plink --range}

\item{rank}{Same as plink --rank}

\item{read_dists}{Same as plink --read-dists}

\item{read_freq}{Same as plink --read-freq}

\item{read_genome}{Same as plink --read-genome}

\item{read_genome_list}{Same as plink --read-genome-list}

\item{read_genome_minimal}{Same as plink --read-genome-minimal}

\item{recessive}{Same as plink --recessive}

\item{recode12}{Same as plink --recode12}

\item{recode}{Same as plink --recode}

\item{recodeA}{Same as plink --recodeA}

\item{recodeAD}{Same as plink --recodeAD}

\item{recodeHV}{Same as plink --recodeHV}

\item{recode_allele}{Same as plink --recode-allele}

\item{recode_beagle}{Same as plink --recode-beagle}

\item{recode_bimbam}{Same as plink --recode-bimbam}

\item{recode_fastphase}{Same as plink --recode-fastphase}

\item{recode_lgen}{Same as plink --recode-lgen}

\item{recode_rlist}{Same as plink --recode-rlist}

\item{recode_structure}{Same as plink --recode-structure}

\item{recode_vcf}{Same as plink --recode-vcf}

\item{recode_whap}{Same as plink --recode-whap}

\item{reference}{Same as plink --reference}

\item{reference_allele}{Same as plink --reference-allele}

\item{regress_distance}{Same as plink --regress-distance}

\item{regress_pcs}{Same as plink --regress-pcs}

\item{regress_rel}{Same as plink --regress-rel}

\item{rel_check}{Same as plink --rel-check}

\item{rel_cutoff}{Same as plink --rel-cutoff}

\item{remove}{Same as plink --remove}

\item{remove_cluster_names}{Same as plink --remove-cluster-names}

\item{remove_clusters}{Same as plink --remove-clusters}

\item{remove_fam}{Same as plink --remove-fam}

\item{rerun}{Same as plink --rerun}

\item{rice}{Same as plink --rice}

\item{sample}{Same as plink --sample}

\item{score}{Same as plink --score}

\item{score_no_mean_imputation}{Same as plink --score-no-mean-imputation}

\item{script}{Same as plink --script}

\item{seed}{Same as plink --seed}

\item{set}{Same as plink --set}

\item{set_by_all}{Same as plink --set-by-all}

\item{set_collapse_all}{Same as plink --set-collapse-all}

\item{set_hh_missing}{Same as plink --set-hh-missing}

\item{set_max}{Same as plink --set-max}

\item{set_me_missing}{Same as plink --set-me-missing}

\item{set_missing_nonsnp_ids}{Same as plink --set-missing-nonsnp-ids}

\item{set_missing_snp_ids}{Same as plink --set-missing-snp-ids}

\item{set_missing_var_ids}{Same as plink --set-missing-var-ids}

\item{set_names}{Same as plink --set-names}

\item{set_p}{Same as plink --set-p}

\item{set_r2}{Same as plink --set-r2}

\item{set_r2_phase}{Same as plink --set-r2-phase}

\item{set_table}{Same as plink --set-table}

\item{set_test}{Same as plink --set-test}

\item{sex}{Same as plink --sex}

\item{sheep}{Same as plink --sheep}

\item{show_tags}{Same as plink --show-tags}

\item{silent}{Same as plink --silent}

\item{simulate}{Same as plink --simulate}

\item{simulate_haps}{Same as plink --simulate-haps}

\item{simulate_label}{Same as plink --simulate-label}

\item{simulate_missing}{Same as plink --simulate-missing}

\item{simulate_n}{Same as plink --simulate-n}

\item{simulate_ncases}{Same as plink --simulate-ncases}

\item{simulate_ncontrols}{Same as plink --simulate-ncontrols}

\item{simulate_prevalence}{Same as plink --simulate-prevalence}

\item{simulate_qt}{Same as plink --simulate-qt}

\item{simulate_tags}{Same as plink --simulate-tags}

\item{snp}{Same as plink --snp}

\item{snps}{Same as plink --snps}

\item{snps_only}{Same as plink --snps-only}

\item{specific_haplotype}{Same as plink --specific-haplotype}

\item{split_x}{Same as plink --split-x}

\item{standard_beta}{Same as plink --standard-beta}

\item{subset}{Same as plink --subset}

\item{swap_parents}{Same as plink --swap-parents}

\item{swap_sibs}{Same as plink --swap-sibs}

\item{swap_unrel}{Same as plink --swap-unrel}

\item{tab}{Same as plink --tab}

\item{tag_kb}{Same as plink --tag-kb}

\item{tag_mode2}{Same as plink --tag-mode2}

\item{tag_r2}{Same as plink --tag-r2}

\item{tail_pheno}{Same as plink --tail-pheno}

\item{tdt}{Same as plink --tdt}

\item{test_all}{Same as plink --test-all}

\item{test_mishap}{Same as plink --test-mishap}

\item{test_missing}{Same as plink --test-missing}

\item{test_snp}{Same as plink --test-snp}

\item{tests}{Same as plink --tests}

\item{tfam}{Same as plink --tfam}

\item{tfile}{Same as plink --tfile}

\item{thin}{Same as plink --thin}

\item{thin_count}{Same as plink --thin-count}

\item{threads}{Same as plink --threads}

\item{to}{Same as plink --to}

\item{to_bp}{Same as plink --to-bp}

\item{to_kb}{Same as plink --to-kb}

\item{to_mb}{Same as plink --to-mb}

\item{tped}{Same as plink --tped}

\item{transpose}{Same as plink --transpose}

\item{trend}{Same as plink --trend}

\item{tucc}{Same as plink --tucc}

\item{twolocus}{Same as plink --twolocus}

\item{unbounded}{Same as plink --unbounded}

\item{unrelated_heritability}{Same as plink --unrelated-heritability}

\item{update_alleles}{Same as plink --update-alleles}

\item{update_chr}{Same as plink --update-chr}

\item{update_cm}{Same as plink --update-cm}

\item{update_ids}{Same as plink --update-ids}

\item{update_map}{Same as plink --update-map}

\item{update_name}{Same as plink --update-name}

\item{update_parents}{Same as plink --update-parents}

\item{update_sex}{Same as plink --update-sex}

\item{vcf}{Same as plink --vcf}

\item{vcf_filter}{Same as plink --vcf-filter}

\item{vcf_half_call}{Same as plink --vcf-half-call}

\item{vcf_idspace_to}{Same as plink --vcf-idspace-to}

\item{vcf_min_qual}{Same as plink --vcf-min-qual}

\item{vegas}{Same as plink --vegas}

\item{version}{Same as plink --version}

\item{vif}{Same as plink --vif}

\item{whap}{Same as plink --whap}

\item{window}{Same as plink --window}

\item{with_freqs}{Same as plink --with-freqs}

\item{with_phenotype}{Same as plink --with-phenotype}

\item{with_reference}{Same as plink --with-reference}

\item{within}{Same as plink --within}

\item{write_cluster}{Same as plink --write-cluster}

\item{write_covar}{Same as plink --write-covar}

\item{write_dosage}{Same as plink --write-dosage}

\item{write_set}{Same as plink --write-set}

\item{write_set_r2}{Same as plink --write-set-r2}

\item{write_snplist}{Same as plink --write-snplist}

\item{xchr_model}{Same as plink --xchr-model}

\item{zero_cluster}{Same as plink --zero-cluster}

\item{zero_cms}{Same as plink --zero-cms}

\item{one}{Same as plink --1}

\item{twothreefile}{Same as plink --23file}

\item{stdout}{Passed to system2, see its documentation.}

\item{stderr}{Passed to system2, see its documentation.}

\item{wait}{Logical. If FALSE, the plink process will fork into the background.}
}
\description{
A wrapper for plink
}

