% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{plTrim}
\alias{plTrim}
\title{Trim plink files}
\usage{
plTrim(pl_gwas, suffix = "trimmed")
}
\arguments{
\item{pl_gwas}{PlGwasC object.}

\item{suffix}{character. Suffix to the new plink file names.}
}
\value{
PlGwasC object
}
\description{
This function calculates number of individuals in .fam file (n1)
and number of individuals in phenotype file (n2). If n1 > n2, then
all the individuals not included in the phenotype file will be 
removed from plink files.
}
\author{
Kaiyin Zhong, Fan Liu
}

