% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{qq}
\alias{qq}
\title{QQ plot of one p-value vector}
\usage{
qq(pvector)
}
\arguments{
\item{pvector}{p-value vector}
}
\value{
A ggplot object
}
\description{
QQ plot of one p-value vector
}
\author{
kaiyin
}

