% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{qq2}
\alias{qq2}
\title{QQ plot of two p-value vector}
\usage{
qq2(p1, p2)
}
\arguments{
\item{p1}{First p-value vector}

\item{p2}{Second p-value vector}
}
\value{
A ggplot object
}
\description{
QQ plot of two p-value vector
}
\author{
kaiyin
}

