% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{qqmulti}
\alias{qqmulti}
\title{QQ plot of multiple p-value vectors}
\usage{
qqmulti(...)
}
\arguments{
\item{...}{p-value vectors. These vectors don't have to have the same length.}
}
\value{
A ggplot object. One QQ plot for each p-value vector and they superposed one after another.
}
\description{
QQ plot of multiple p-value vectors
}
\author{
kaiyin
}

