% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{randNormDat}
\alias{randNormDat}
\title{Generate a m by n data.frame from normal distribution}
\usage{
randNormDat(m, n)
}
\arguments{
\item{m}{integer. Number of rows.}

\item{n}{integer. Number of columns.}
}
\description{
Generate a m by n data.frame from normal distribution
}
\author{
Kaiyin Zhong
}

