% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{randomString}
\alias{randomString}
\title{Generate a single alpha-numeric random string}
\usage{
randomString(string_length = 6)
}
\arguments{
\item{string_length}{integer.}
}
\value{
character.
}
\description{
Generate a single alpha-numeric random string
}
\author{
Kaiyin Zhong, Fan Liu
}

