% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{randomStrings}
\alias{randomStrings}
\title{Generate random strings}
\usage{
randomStrings(n, string_length = 6)
}
\arguments{
\item{n}{integer. Number of string to generate.}

\item{string_length}{integer. Length of each string.}
}
\value{
character.
}
\description{
Generate random strings
}
\author{
Kaiyin Zhong, Fan Liu
}

