% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{rbedInfo}
\alias{rbedInfo}
\title{Constructor of RbedInfoC class}
\usage{
rbedInfo(bedstem, db_setup = FALSE)
}
\arguments{
\item{bedstem}{character. Path to bed file without extension.}

\item{db_setup}{logical. Whether to setup SQLite database for .bim, .fam and .frq files.}
}
\value{
An RbedInfoC object.
}
\description{
Constructor of RbedInfoC class
}
\author{
Kaiyin Zhong, Fan Liu
}

