% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\docType{class}
\name{RbedInfoC-class}
\alias{.RbedInfoC}
\alias{RbedInfoC-class}
\title{S4 class for necessary info to read a bed file into R}
\description{
S4 class for necessary info to read a bed file into R
}
\section{Slots}{

\describe{
\item{\code{pl_info}}{PlInfoC object}

\item{\code{jbed}}{jobjRef object, of Bed class in java}

\item{\code{nsnp}}{numeric. Number of SNPs.}

\item{\code{nindiv}}{numeric. Number of individuals.}

\item{\code{nindiv_appr}}{numeric. Apparent number of individuals.}

\item{\code{bytes_snp}}{numeric. Number of bytes used for each SNP.}
}}

