% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{read.phe.table}
\alias{read.phe.table}
\title{Read phenotype file}
\usage{
read.phe.table(file)
}
\arguments{
\item{file}{character, path to phenotype file.}
}
\value{
data.frame
}
\description{
Read phenotype file
}
\author{
kaiyin
}

