% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readAssoc}
\alias{readAssoc}
\title{Read PLINK .assoc files}
\usage{
readAssoc(filename, cn_select = collenv$.assoc_header)
}
\arguments{
\item{filename}{character. Filename}

\item{cn_select}{character. Columns to read.}
}
\value{
data.frame
}
\description{
Read PLINK .assoc files
}
\author{
Kaiyin Zhong
}

