% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readBim}
\alias{readBim}
\title{Read plink .bim files}
\usage{
readBim(filename, cn_select = "..all")
}
\arguments{
\item{filename}{.bim file path}

\item{cn_select}{a character vector for selected colnames}
}
\value{
a data.frame
}
\description{
Read plink .bim files
}
\author{
Kaiyin Zhong, Fan Liu
}

