% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{readBmBin}
\alias{readBmBin}
\title{Read columns into an R matrix from a big.matrix .bin file}
\usage{
readBmBin(bin_file, ncols_to_read)
}
\arguments{
\item{bin_file}{character. Path to .bin file}

\item{ncols_to_read}{integer.}
}
\value{
matrix
}
\description{
Read columns into an R matrix from a big.matrix .bin file
}
\author{
Kaiyin Zhong, Fan Liu
}

