% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readcoll}
\alias{readcoll}
\title{Read PLINK assoc files int matrices and store them in the QCDH data hub}
\usage{
readcoll(hub, plinkOutExt)
}
\arguments{
\item{hub}{QCDH data hub}

\item{plinkOutExt}{Extension name of the PLINK output, e.g. ".assoc.linear"}
}
\description{
Read PLINK assoc files int matrices and store them in the QCDH data hub
}

