% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{readDesc}
\alias{readDesc}
\title{Read big.matrix .desc file}
\usage{
readDesc(desc_filename)
}
\arguments{
\item{desc_filename}{character. Path to .desc file}
}
\value{
description object
}
\description{
Read big.matrix .desc file
}
\author{
Kaiyin Zhong, Fan Liu
}

