% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readFam}
\alias{readFam}
\title{Read plink .fam files}
\usage{
readFam(filename, cn_select = "..all")
}
\arguments{
\item{filename}{.fam file path}

\item{cn_select}{a character vector for selected colnames}
}
\value{
a data.frame
}
\description{
Read plink .fam files
}
\examples{
\dontrun{
bim = readBim("mmp13.bim")
bim1 = readBim("mmp13.bim", "..all")
fam = readFam("mmp13.fam", "..all")
}

}
\author{
Kaiyin Zhong, Fan Liu
}

