% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readFunFactory}
\alias{readFunFactory}
\title{Generate read_fun for ReadInfo class}
\usage{
readFunFactory(header)
}
\arguments{
\item{header}{logical. Whether the input file has a header line.}
}
\value{
function.
}
\description{
Generate read_fun for ReadInfo class
}
\author{
Kaiyin Zhong, Fan Liu
}

