% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{readGwasOut}
\alias{readGwasOut}
\title{Read GWAS output from plink
If the GWAS is finished, returns a data.frame, 
otherwise returns NULL.}
\usage{
readGwasOut(pl_gwas, cn_select = "..all", rmGwasOut = TRUE)
}
\arguments{
\item{pl_gwas}{PlGwasC object.}

\item{cn_select}{Colnames to select. Default to "..all"}

\item{rmGwasOut}{Logical. Whether to remove GWAS output files after finished reading them. Default to TRUE.}
}
\value{
data.frame
}
\description{
Read GWAS output from plink
If the GWAS is finished, returns a data.frame, 
otherwise returns NULL.
}
\author{
Kaiyin Zhong, Fan Liu
}

