% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\docType{class}
\name{ReadInfo-class}
\alias{.ReadInfo}
\alias{ReadInfo-class}
\title{An S4 class to represent information about a whitespace-delimited text file to be read into R}
\description{
An S4 class to represent information about a whitespace-delimited text file to be read into R
}
\section{Slots}{

\describe{
\item{\code{filename}}{Path of the file}

\item{\code{cnames}}{character vector of column names}

\item{\code{header}}{logical. Whether the first line is header}

\item{\code{read_fun}}{function. The function to be used when reading this file}
}}

