% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\docType{methods}
\name{readInfo}
\alias{readInfo}
\alias{readInfo,character,character-method}
\alias{readInfo,character,missing-method}
\title{ReadInfo constructor}
\usage{
readInfo(filename, cnames)

\S4method{readInfo}{character,missing}(filename)

\S4method{readInfo}{character,character}(filename, cnames)
}
\arguments{
\item{filename}{Path of the file to read}

\item{cnames}{character. Expected column names (header).}
}
\value{
ReadInfo object
}
\description{
This function takes a file path as parameter, assuming the file is whitspace delimited, 
not quoted, and has a header line. It returns a ReadInfo object.
}
\examples{
\dontrun{
ri = readInfo("mmp13.frq")
ri@cnames 
ri@filename
ri@header 
}
}
\author{
Kaiyin Zhong, Fan Liu
}

