% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readLogistic}
\alias{readLogistic}
\title{Read PLINK logistic regression output files.}
\usage{
readLogistic(filename, cn_select = collenv$.linear_header)
}
\arguments{
\item{filename}{character. Filename.}

\item{cn_select}{character. Columns to read.}
}
\value{
data.frame
}
\description{
Read PLINK logistic regression output files.
}
\author{
Kaiyin Zhong
}

