% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{readPhe}
\alias{readPhe}
\title{Read phenotype file}
\usage{
readPhe(pl_gwas, cn_select = "..all")
}
\arguments{
\item{pl_gwas}{PlGwasC object}

\item{cn_select}{Colnames to select. Default to "..all", which means all columns are read in.}
}
\value{
data.frame
}
\description{
Read phenotype file
}
\author{
Kaiyin Zhong, Fan Liu
}

