% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readQassoc}
\alias{readQassoc}
\title{Read .qassoc files}
\usage{
readQassoc(filename, cn_select = collenv$.qassoc_header)
}
\arguments{
\item{filename}{Path of the file to read}

\item{cn_select}{a character vector for selected colnames}
}
\value{
data.frame.
}
\description{
Read .qassoc files
}
\author{
Kaiyin Zhong
}

