% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{realBedSize}
\alias{realBedSize}
\title{File size of bed file}
\usage{
realBedSize(rbed_info)
}
\arguments{
\item{rbed_info}{RbedInfoC object}
}
\value{
numeric. Size of bed file.
}
\description{
File size of bed file
}
\author{
Kaiyin Zhong, Fan Liu
}

