% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{removeTag}
\alias{removeTag}
\title{Remove GWAS results by tag}
\usage{
removeTag(x, type = "gwas")
}
\arguments{
\item{x}{character. Tag name.}

\item{type}{character. Type of tag.}
}
\description{
Remove GWAS results by tag
}
\author{
Kaiyin Zhong, Fan Liu
}

