% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rep_dat}
\alias{rep_dat}
\title{Batch version of the \code{rep_models} function:}
\usage{
rep_dat(mat, snpdat, pheno_name, covar_names)
}
\arguments{
\item{mat}{A data frame containing a collection of single SNP genotypes, SNP names should be column names.}

\item{snpdat}{A data frame containing pairs of SNP names. Each row should contain a one pair.}

\item{pheno_name}{Character. Name of phenotype.}

\item{covar_names}{Character. Vector of names of covariates.}
}
\value{
A data frame with five columns: snp1, snp2, p1 (p value of snp1), p2 (pvalue of snp2) and pcol (p value of collapsed genotype).
}
\description{
Batch version of the \code{rep_models} function:
}

