% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rep_models}
\alias{rep_models}
\title{Given a pair of SNPs, fit a model using single and collapsed genotypes}
\usage{
rep_models(mat, snp1, snp2, pheno_name, covar_names)
}
\arguments{
\item{mat}{A data frame containing a collection of single SNP genotypes, SNP names should be column names.}

\item{snp1}{Character.Name of first SNP}

\item{snp2}{Character.Name of second SNP}

\item{pheno_name}{Character. Name of phenotype.}

\item{covar_names}{Character. Vector of names of covariates.}
}
\value{
A data frame with five columns: snp1, snp2, p1 (p value of snp1), p2 (pvalue of snp2) and pcol (p value of collapsed genotype).
}
\description{
Given a pair of SNPs, fit a model using single and collapsed genotypes
}

