% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{reprClasses}
\alias{reprClasses}
\title{Represent classes of a data.frame in a character vector}
\usage{
reprClasses(dat)
}
\arguments{
\item{dat}{data.frame}
}
\value{
character vector
}
\description{
Represent classes of a data.frame in a character vector
}
\examples{
\dontrun{
dat = randNormDat(4, 2)
x = capture.output(reprClasses(dat), file = NULL)
x = eval(parse(text = x))
all(x == colClasses(dat))
}

}
\author{
Kaiyin Zhong, Fan Liu
}

