% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{rmFilesByStem}
\alias{rmFilesByStem}
\title{Remove files by matching the starting part}
\usage{
rmFilesByStem(x)
}
\arguments{
\item{x}{character, PlInfoC, or RbedInfoC object.}
}
\description{
If \code{x} is a string, then this function matches \code{x*} by globbing.
If \code{x} is a "PlInfoC" object, it matches \code{x@plink_stem*},
If \code{x} is a "RbedInfoC" object, it matches \code{x@pl_info@plink_stem*}.
Otherwise nothing is removed.
}
\author{
Kaiyin Zhong, Fan Liu
}

