% Generated by roxygen2 (4.0.2): do not edit by hand
\name{routine1}
\alias{routine1}
\title{QCDH routine 1}
\usage{
routine1(wDir = ".", taskName, plinkArgs, initGwas = FALSE, initGwasArgs,
  pFilter = 0.05, nMaxShift)
}
\arguments{
\item{wDir}{Working directory. All paths within this function will be relative to this.}

\item{taskName}{Name of the QCDH task.}

\item{plinkArgs}{Arguments to be passed to PLINK (in QCDH analysis)}

\item{initGwas}{Whether to perform an initial GWAS}

\item{initGwasArgs}{Arguments to pass to PLINK for the initial GWAS}

\item{pFilter}{Filter out SNPs with p value higher than this (in the initial GWAS)}

\item{nMaxShift}{Maximum shift number (for genotype collapsing)}
}
\value{
hubtask An environment containing info and results from the current task.
}
\description{
This routine collects info on the working directory, performs an
initial GWAS as a basis for SNP selection by p values, generates
bed files with collapsed genotypes, analyzes them, and saves intermediate
and final results in an environment named after the task.
This routine is called by \code{qcdhRoutine}.
}

