% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{runGwas}
\alias{runGwas}
\title{Run a GWAS}
\usage{
runGwas(pl_gwas, wait = TRUE, save_pl_gwas = FALSE)
}
\arguments{
\item{pl_gwas}{PlGwasC object}

\item{wait}{logical. Wait until GWAS is finished if this is set to TRUE. Default to FALSE.}

\item{save_pl_gwas}{logical. Whether to save the plGwas object. Default to FALSE.}
}
\description{
Run a GWAS
}
\author{
Kaiyin Zhong, Fan Liu
}

