% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{saveDesc}
\alias{saveDesc}
\title{Save big.matrix description object to disk}
\usage{
saveDesc(desc_obj, desc_filename)
}
\arguments{
\item{desc_obj}{big.matrix description object}

\item{desc_filename}{character. Output file description file path.}
}
\description{
Binary format is used exclusively.
}
\author{
Kaiyin Zhong, Fan Liu
}

