% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_sqlUtils.R
\name{sendQuery}
\alias{sendQuery}
\title{Send query to SQLite database}
\usage{
sendQuery(db_name, query_string)
}
\arguments{
\item{db_name}{character. Path to database.}

\item{query_string}{character. Query string.}
}
\description{
Send query to SQLite database
}
\author{
Kaiyin Zhong, Fan Liu
}

