% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{setOptModel}
\alias{setOptModel}
\title{Set analysis model}
\usage{
setOptModel(pl_gwas, mod = "linear")
}
\arguments{
\item{pl_gwas}{PlGwasC object.}

\item{mod}{character. One of "linear", "logistic" or "assoc", default to "linear".}
}
\value{
PlGwasC object
}
\description{
Set analysis model
}
\author{
Kaiyin Zhong, Fan Liu
}

