% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bedinfo.R
\name{setup}
\alias{setup}
\title{Setup up a directory containing plink files}
\usage{
setup(pl_info)
}
\arguments{
\item{pl_info}{PlInfoC object}
}
\description{
Setup up a directory containing plink files
}
\author{
Kaiyin Zhong, Fan Liu
}

