% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{setupRbed}
\alias{setupRbed}
\title{Setup an RbedInfoC object}
\usage{
setupRbed(rbed_info)
}
\arguments{
\item{rbed_info}{RbedInfoC object}
}
\value{
RbedInfoC object
}
\description{
The setup job includes the following tasks: 1. Set up the PlInfoC object.
2. Calculate number of bytes used by each SNP. 3. Calculate the Number of individuals.
4. Calculate total number of SNPs. 5. Validate the RbedInfoC object.
}
\author{
Kaiyin Zhong, Fan Liu
}

