% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shiftDataByCol}
\alias{shiftDataByCol}
\title{Shift columns of a matrix}
\usage{
shiftDataByCol(dat)
}
\arguments{
\item{dat}{The matrix to be shifted.}

\item{The}{shifted matrix.}
}
\description{
Column 1 is kept the same, column 2 is shifted upwards by 1 row, column 3 by 2, etc.
See \url{https://gist.github.com/kindlychung/757f816f5b2f9e8d36b8#file-shiftdatabycol_example-txt}
}

