% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{shiftedStem}
\alias{shiftedStem}
\title{Add a "shift" suffix to a stem}
\usage{
shiftedStem(stem, n_shift)
}
\arguments{
\item{stem}{character.}

\item{n_shift}{numeric.}
}
\value{
character.
}
\description{
Add a "shift" suffix to a stem
}
\examples{
\dontrun{
# add suffix to stem
shiftedStem("a", 100) == "a_shift_0100"
shiftedStem("home/a", 100) == "home/a_shift_0100"
shiftedStem("/home/a", 100) == "/home/a_shift_0100"
shiftedStem(c("/home/a", "/home/b"), 100) == c("/home/a_shift_0100",
		"/home/b_shift_0100")
}

}
\author{
Kaiyin Zhong, Fan Liu
}

