% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{slurp}
\alias{slurp}
\title{Read a text file into a single string}
\usage{
slurp(filename)
}
\arguments{
\item{filename}{character. Input filename.}
}
\value{
character
}
\description{
Read a text file into a single string
}
\author{
Kaiyin Zhong, Fan Liu
}

