% Generated by roxygen2 (4.0.2): do not edit by hand
\name{snpPairLookin.quant}
\alias{snpPairLookin.quant}
\title{Look into a pairs of SNPs}
\usage{
snpPairLookin.quant(g1, g2, pheno)
}
\arguments{
\item{g1}{Numeric. A vector of genotypes for SNP 1}

\item{g2}{Numeric. A vector of genotypes for SNP 2}

\item{pheno}{Numeric. A vector of phenotypes.}
}
\value{
A ggplot object.
}
\description{
Once your got a pair of SNPs that looks interesting from your GWAS, you can use this
function to look into their relations with the (quantitative) phenotype.
}

