% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpPos.R
\name{snpPos}
\alias{snpPos}
\title{Retrive SNP positions from UCSU database}
\usage{
snpPos(snps, rm_underscore = TRUE, ref = c("hg18", "hg19"),
  snpdb = c("snp138", "snp137"))
}
\arguments{
\item{snps}{A vector of SNP names}

\item{rm_underscore}{Remove irregular chromosome names}

\item{ref}{Either "hg18" or "hg19"}

\item{snpdb}{Either "snp138" or "snp137"}
}
\value{
A data frame containing positions of given SNPs
}
\description{
Retrive SNP positions from UCSU database
}
\author{
kaiyin
}

