% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_strConcat.R
\name{snpPosSNP138}
\alias{snpPosSNP138}
\title{Retrive SNP positions from hg19 database}
\usage{
snpPosSNP138(snps, rm_underscore = TRUE)
}
\arguments{
\item{snps}{A vector of SNP names}

\item{rm_underscore}{Remove irregular chromosome names}
}
\value{
data.frame
}
\description{
Retrive SNP positions from hg19 database
}
\author{
kaiyin
}

