% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{snpRowId}
\alias{snpRowId}
\title{Get row number of SNPs from their names}
\usage{
snpRowId(pl_info, snp_names)
}
\arguments{
\item{pl_info}{PlInfoC object.}

\item{snp_names}{character. Vector of SNP names.}
}
\value{
integer. Vector of row numbers.
}
\description{
Get row number of SNPs from their names
}
\author{
Kaiyin Zhong, Fan Liu
}

