% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bigmatrixUtils.R
\name{spit}
\alias{spit}
\title{Write strings to a file}
\usage{
spit(s, filename)
}
\arguments{
\item{s}{character. Strings to write.}

\item{filename}{character. Path to output file.}
}
\description{
Write strings to a file
}
\author{
Kaiyin Zhong, Fan Liu
}

