% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_bedinfo.R
\name{sqliteFilePl}
\alias{sqliteFilePl}
\title{SQLite file of a PlInfoC object}
\usage{
sqliteFilePl(x)
}
\arguments{
\item{x}{PlInfoC or PlGwasC object}
}
\value{
character. Path to SQLite database file.
}
\description{
SQLite file of a PlInfoC object
}
\author{
Kaiyin Zhong, Fan Liu
}

