% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{stopFormat}
\alias{stopFormat}
\title{Stop with format string}
\usage{
stopFormat(...)
}
\arguments{
\item{...}{passed to \code{sprintf}}
}
\description{
Stop with format string
}
\author{
Kaiyin Zhong, Fan Liu
}

