% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_strConcat.R
\docType{methods}
\name{strConcat}
\alias{strConcat}
\title{Concatenate a vector of strings}
\usage{
strConcat(ss, sep = "")
}
\arguments{
\item{ss}{vector of strings}

\item{sep}{a length-1 string used as separator, default to ""}
}
\value{
a string
}
\description{
Concatenate a vector of strings
}
\examples{
\dontrun{
strConcat(letters)
strConcat(letters, " ")
}

}
\author{
Kaiyin Zhong, Fan Liu
}

