% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_strConcat.R
\name{strVectorSQLRepr}
\alias{strVectorSQLRepr}
\title{String representation of a character vector for SQLite consumption}
\usage{
strVectorSQLRepr(vec, print_out = FALSE, single_quote = TRUE)
}
\arguments{
\item{vec}{character.}

\item{print_out}{logical. Print out the string representation when set to TRUE.}

\item{single_quote}{logical. Whether to use single quote for each element. Use double quote if set to FALSE. Default to TRUE.}
}
\description{
Transform a character vector (e.g. \code{c("a", "b")} into a string representation 
that can be used in a SQLite query (e.g. "('a', 'b')").
}
\author{
Kaiyin Zhong
}

