% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{systemFormat}
\alias{systemFormat}
\title{Call system command with format string}
\usage{
systemFormat(...)
}
\arguments{
\item{...}{passed to \code{sprintf}}
}
\description{
Call system command with format string
}
\author{
Kaiyin Zhong, Fan Liu
}

