% Generated by roxygen2 (4.0.2): do not edit by hand
\name{taskAnalyze}
\alias{taskAnalyze}
\title{Analyze shifted bed files}
\usage{
taskAnalyze(hub, taskName)
}
\arguments{
\item{hub}{Environment. QCDH data hub}

\item{taskName}{Character. Name of the task.}
}
\description{
Analyze shifted bed files
}

