% Generated by roxygen2 (4.0.2): do not edit by hand
\name{taskBedsPlinkOut}
\alias{taskBedsPlinkOut}
\title{Prepare shifted bed files for current task.}
\usage{
taskBedsPlinkOut(hub, taskName, plinkOutFile, pvalThresh = 0.05,
  nMinShift = 1, nMaxShift)
}
\arguments{
\item{hub}{Environment. QCDH data hub.}

\item{taskName}{Character. Name of the task.}

\item{plinkOutFile}{Character. PLINK assoc file of the initial GWAS.}

\item{pvalThresh}{Numeric. p value threshold for extraction of SNPs.}
}
\description{
Read p values from the initial GWAS, get a list of SNPs by the given p value threshold,
extract the genotypes of these SNPs and create a new bed file, then generate shifted bed files
from this new bed file.
}

