% Generated by roxygen2 (4.0.2): do not edit by hand
\name{taskinfo}
\alias{taskinfo}
\title{Prepare paths and filenames for the task}
\usage{
taskinfo(hub, taskName, plinkArgs, initGwasArgs, initGwas = FALSE)
}
\arguments{
\item{hub}{Environment. QCDH data hub.}

\item{taskName}{Character. Name of the task.}

\item{plinkArgs}{List. Arguments to be passed to PLINK}

\item{initGwasArgs}{List. Arguments to be passed to PLINK}

\item{initGwas}{Logical. Whether to perform a initial GWAS.}
}
\description{
A hepler function for sorting out folder and file names for plots, shifted bed files etc.
It also figures out extension names of PLINK output.
}

