% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_rbedInfo.R
\name{theoBedSize}
\alias{theoBedSize}
\title{Theoretical size of bed file}
\usage{
theoBedSize(rbed_info)
}
\arguments{
\item{rbed_info}{RbedInfoC object}
}
\value{
numeric. Theoretical size of bed file.
}
\description{
Computed from dimensions of bim an fam files.
}
\author{
Kaiyin Zhong, Fan Liu
}

