% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateMinPvals}
\alias{updateMinPvals}
\title{Calculate minial p values for QCDH}
\usage{
updateMinPvals(hub, bpdiff = 5e+05)
}
\arguments{
\item{hub}{QCDH data hub}

\item{bpdiff}{Numeric. Upper bound of base position distance between a pair of SNPs.}
}
\description{
This function calculates minimal p values for each collection of SNP pairs that share the same first SNP.
It imposes two conditions: (1) Each pair should be in the same chromosome. (2) The distance between each pair of
SNPs should be below a certain threshold (specified by user).
}

