% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_permute.R
\name{validPhe}
\alias{validPhe}
\title{Validate a phenotype file}
\usage{
validPhe(phe_file, ...)
}
\arguments{
\item{phe_file}{character. Phenotype file.}

\item{...}{Passed to read.table}
}
\value{
FALSE when the file is invalid, or a data.frame when it is.
}
\description{
Validate a phenotype file
}
\author{
Kaiyin Zhong, Fan Liu
}

