% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{write.phe.table}
\alias{write.phe.table}
\title{Write a phenotype data.frame to file}
\usage{
write.phe.table(phe, file)
}
\arguments{
\item{phe}{data.frame}

\item{file}{character, path to phenotype file.}
}
\description{
Write a phenotype data.frame to file
}
\author{
Kaiyin Zhong
}

