% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_plGwas.R
\name{binPhe}
\alias{binPhe}
\title{Check whether phenotype of a GWAS is binary}
\usage{
binPhe(pl_gwas, na_value = c(-9, 0))
}
\arguments{
\item{pl_gwas}{PlGwasC object.}

\item{na_value}{A vector of codes that represent missing values.}
}
\value{
logical
}
\description{
Check whether phenotype of a GWAS is binary
}
\author{
Kaiyin Zhong, Fan Liu
}

