% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils.R
\name{collapse}
\alias{collapse}
\title{Collpase genotypes}
\usage{
collapse(g1, g2, collapse_matrix = NULL)
}
\arguments{
\item{g1}{numeric, genotype vector 1.}

\item{g2}{numeric, genotype vector 2.}

\item{collapse_matrix}{matrix of integers range from 0 to 3.}
}
\value{
numeric, collapsed genotype of g1 and g2.
}
\description{
Collpase genotypes
}
\author{
Kaiyin Zhong
}

