% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_manhattan.R
\name{contrastPlot}
\alias{contrastPlot}
\title{Produce contrast Manhattan plot}
\usage{
contrastPlot(chr, bp, p, gcdh_p, snp, ...)
}
\arguments{
\item{chr}{integer. Chromosome vector.}

\item{bp}{integer. Position vector.}

\item{p}{numeric. P-value vector.}

\item{gcdh_p}{numeric. GCDH p-value vector.}

\item{snp}{character. SNP name vector.}

\item{...}{passed to \code{manhattanPlot}}
}
\value{
ggplot object.
}
\description{
Overlays p-values from single-SNP method and GCDH.
}
\author{
Kaiyin Zhong
}

